Internals

Everything that is not exported is considered internal.

PAndQ.union_typeof — Function

union_typeof

Operators

Union Types

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of LogicalOperators that take zero arguments.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of LogicalOperators that take one argument.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of LogicalOperators that take two arguments.

PAndQ.LogicalOperator — Type

LogicalOperator

The Union of logical operators.

PAndQ.CommutativeOperator — Type

CommutativeOperator

The Union of LogicalOperators with the commutative property.

PAndQ.AssociativeOperator — Type

AssociativeOperator

The Union of LogicalOperators with the associative property.

PAndQ.LeftNeutralOperator — Type

LeftNeutralOperator

The Union of LogicalOperators that have one or more left_neutrals.

PAndQ.RightNeutralOperator — Type

RightNeutralOperator

The Union of LogicalOperators that have one or more right_neutrals.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

Union Types

PAndQ.AtomicProposition — Type

AtomicProposition

A Proposition that is known by its type to be logically equivalent to an Atom.

PAndQ.LiteralProposition — Type

LiteralProposition

A Proposition that is known by its type to be logically equivalent to a Literal.

AbstractTrees.jl

AbstractTrees.children — Function

children(::Proposition)

Return an iterator over the child nodes of the given Proposition.

Examples

julia> @atomize PAndQ.children(p)
()

julia> @atomize PAndQ.children(¬p)
(Variable(:p),)

julia> @atomize PAndQ.children(p ∧ q)
2-element Vector{Tree{typeof(identity), Variable}}:
 p
 q

AbstractTrees.nodevalue — Function

nodevalue(::Compound)

Examples

julia> @atomize PAndQ.nodevalue(¬p)
not (generic function with 19 methods)

julia> @atomize PAndQ.nodevalue(p ∧ q)
and (generic function with 18 methods)

AbstractTrees.printnode — Function

printnode(::IO, ::Proposition; kwargs...)

See also Proposition.

Examples

julia> @atomize PAndQ.printnode(stdout, p)
p
julia> @atomize PAndQ.printnode(stdout, ¬p)
¬
julia> @atomize PAndQ.printnode(stdout, p ∧ q)
∧

Utility

PAndQ.child — Function

child(x)

Equivalent to only(children(x))

PAndQ.union_all_type — Function

union_all_type(::Proposition)

Return the UnionAll type of a Proposition.

Examples

julia> @atomize PAndQ.union_all_type(p)
Variable

julia> @atomize PAndQ.union_all_type(p ∧ q)
Tree

PAndQ.symbol_value — Function

symbol_value

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as a Variable if it is undefined in the caller's scope. If isexpr(x, :$), return an expression that instantiates it as a Constant. If x is another expression, traverse it with recursive calls to atomize Otherise, return x.

Printing

PAndQ.symbol_of — Function

symbol_of(::LogicalOperator)

Return the Unicode character that is an alias for the given LogicalOperator.

Examples

julia> PAndQ.symbol_of(⊤)
:⊤

julia> PAndQ.symbol_of(¬)
:¬

julia> PAndQ.symbol_of(∧)
:∧

PAndQ.merge_string — Function

merge_string(cell)

PAndQ.parenthesize — Function

parenthesize(::IO, x)

PAndQ.print_node — Function

print_node(io, p)

PAndQ.show_atom — Function

show_atom

Semantics

PAndQ.neutral_operator — Function

neutral_operator(::NullaryOperator)

Return a subtype of AndOr that is the neutral element of the given NullaryOperator.

See also left_neutrals and right_neutrals.

Examples

julia> PAndQ.neutral_operator(⊤)
and (generic function with 18 methods)

julia> PAndQ.neutral_operator(⊥)
or (generic function with 18 methods)

PAndQ.eval_doubles — Function

eval_doubles(f, doubles)