Abstract Types

PAQ.Language — Type

Language

Set of well-formed logical formulae.

Calling an instance of Language will return a vector of valid interpretations.

Supertype of Primitive, Compound, and Truth. ```

PAQ.Compound — Type

Compound <: Language

Compound proposition.

Subtype of Language. Supertype of Propositional.

PAQ.Operator — Type

Operator

Set of functions that operate on a logical Language.

Supertype of Boolean.