Operators

Typing symbols with tab completion is performed by typing \, followed by the given characters, and then the [TAB] key. For example, ⊥ is typed with \bot[TAB]. See also Julia's documentation on Tab Completion and Unicode Input.

Operator associativity determines how operators with the same precedence group their operands. For example, ∧ is left associative. Therefore, p ∧ q ∧ r is equivalent to (p ∧ q) ∧ r. Operator precedence determines how expressions with distinct operators are grouped together. Higher precedence operators will group their operands before lower precedence operators. For example, ∧ has a higher precedence than ∨. Therefore, p ∨ q ∧ r is equivalent to p ∨ (q ∧ r), even though both operators are left associative. See also Julia's documentation on Operator Precedence and Associativity.

Info

== has a precedence of 7, which is higher than that of several binary operators. For some cases, you may need to use parentheses. For example, @p p → q == r will error, but @p (p → q) == r will correctly return false.

Name	Symbol	Tab Completion	Associativity	Precedence

tautology	⊤	\top	none	0

contradiction	⊥	\bot	none	0

not	¬	\neg	right	0

and	∧	\wedge	left	12

nand	⊼	\nand	left	12

nor	⊽	\nor	left	11

or	∨	\vee	left	11

xor	⊻	\xor	left	11

xnor	↔	\leftrightarrow	right	4

imply	→	\rightarrow	right	4

not_imply	↛	\nrightarrow	right	4

converse_imply	←	\leftarrow	right	4

not_converse_imply	↚	\nleftarrow	right	4

Every possible truth table can be constructed with the functionally complete set of operators not and and. For convenience, all sixteen of them have been prepared. There are ten binary operators, with the remaining six being expressed with truth values, individual propositions, and the unary not operator.

julia> @p TruthTable(⊤, ⊥, ¬p, ¬q)
┌────────┬────────┬──────┬──────┬─────────┬─────────┐
│ ⊤      │ ⊥      │ p    │ q    │ ¬p      │ ¬q      │
│ Clause │ Clause │ Atom │ Atom │ Literal │ Literal │
├────────┼────────┼──────┼──────┼─────────┼─────────┤
│ ⊤      │ ⊥      │ ⊤    │ ⊤    │ ⊥       │ ⊥       │
│ ⊤      │ ⊥      │ ⊥    │ ⊤    │ ⊤       │ ⊥       │
├────────┼────────┼──────┼──────┼─────────┼─────────┤
│ ⊤      │ ⊥      │ ⊤    │ ⊥    │ ⊥       │ ⊤       │
│ ⊤      │ ⊥      │ ⊥    │ ⊥    │ ⊤       │ ⊤       │
└────────┴────────┴──────┴──────┴─────────┴─────────┘

Nullary Operators

PAndQ.tautology — Function

tautology()
⊤()

Logical true operator.

⊤ can be typed by \top<tab>.

Examples

julia> ⊤()
tautology (generic function with 1 method)

julia> @p TruthTable(⊤)
┌────────┐
│ ⊤      │
│ Clause │
├────────┤
│ ⊤      │
└────────┘

PAndQ.contradiction — Function

contradiction()
⊥()

Logical false operator.

⊥ can be typed by \bot<tab>.

Examples

julia> ⊥()
contradiction (generic function with 1 method)

julia> @p TruthTable(⊥)
┌────────┐
│ ⊥      │
│ Clause │
├────────┤
│ ⊥      │
└────────┘

Unary Operators

Base.identity — Function

identity(p)

Logical identity operator.

Examples

julia> @p TruthTable(p)
┌──────┐
│ p    │
│ Atom │
├──────┤
│ ⊤    │
│ ⊥    │
└──────┘

PAndQ.not — Function

not(p)
¬p

Logical negation operator.

¬ can be typed by \neg<tab>.

Examples

julia> @p TruthTable(¬p)
┌──────┬─────────┐
│ p    │ ¬p      │
│ Atom │ Literal │
├──────┼─────────┤
│ ⊤    │ ⊥       │
│ ⊥    │ ⊤       │
└──────┴─────────┘

Binary Operators

PAndQ.and — Function

and(p, q)
p ∧ q

Logical conjunction operator.

∧ can be typed by \wedge<tab>.

Examples

julia> @p TruthTable(p ∧ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ∧ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

Base.nand — Function

nand(p, q)
p ⊼ q

Logical non-conjunction operator.

⊼ can be typed by \nand<tab>.

Examples

julia> @p TruthTable(p ⊼ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ⊼ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

Base.nor — Function

nor(p, q)
p ⊽ q

Logical non-disjunction operator.

⊽ can be typed by \nor<tab>.

Examples

julia> @p TruthTable(p ⊽ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ⊽ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAndQ.or — Function

or(p, q)
p ∨ q

Logical disjunction operator.

∨ can be typed by \vee<tab>.

Examples

julia> @p TruthTable(p ∨ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ∨ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

Base.xor — Function

xor(p, q)
p ⊻ q

Logical exclusive disjunction operator.

⊻ can be typed by \xor<tab>.

Examples

julia> @p TruthTable(p ⊻ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ⊻ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

PAndQ.xnor — Function

xnor(p, q)
p ↔ q

Logical exclusive non-disjunction and biconditional operator.

↔ can be typed by \leftrightarrow<tab>.

Examples

julia> @p TruthTable(p ↔ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ↔ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAndQ.imply — Function

imply(p, q)
p → q

Logical implication operator.

→ can be typed by \rightarrow<tab>.

Examples

julia> @p TruthTable(p → q)
┌──────┬──────┬───────┐
│ p    │ q    │ p → q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAndQ.not_imply — Function

not_imply(p, q)
p ↛ q

Logical non-implication operator.

↛ can be typed by \nrightarrow<tab>.

Examples

julia> @p TruthTable(p ↛ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ↛ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

PAndQ.converse_imply — Function

converse_imply(p, q)
p ← q

Logical converse implication operator.

← can be typed by \leftarrow<tab>.

Examples

julia> @p TruthTable(p ← q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ← q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAndQ.not_converse_imply — Function

not_converse_imply(p, q)
p ↚ q

Logical converse non-implication operator.

↚ can be typed by \nleftarrow<tab>.

Examples

julia> @p TruthTable(p ↚ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ↚ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

Reductions

PAndQ.conjunction — Function

conjunction(ps)
⋀(ps)

Equivalent to foldl(and, ps; init = tautology).

⋀ can be typed by \bigwedge<tab>.

See also and and tautology.

Examples

julia> @p ⋀([p, q, r, s])
((p ∧ q) ∧ r) ∧ s

PAndQ.disjunction — Function

disjunction(ps)
⋁(ps)

Equivalent to foldl(or, ps; init = contradiction).

⋁ can be typed by \bigvee<tab>.

See also or and contradiction.

Examples

julia> @p ⋁([p, q, r, s])
((p ∨ q) ∨ r) ∨ s

Base.mapfoldl — Function

mapfoldl(f, lio::LeftNeutralOperator, ps)

Equivalent to mapfoldl(f, lio, ps; init = only(left_neutrals(lio))).

See also LeftNeutralOperator.

Tip

This also works with foldl(lio, ps).

Examples

julia> @p mapfoldl(not, and, [p, q, r, s])
((¬p ∧ ¬q) ∧ ¬r) ∧ ¬s

julia> foldl(and, [])
tautology (generic function with 1 method)

Base.mapfoldr — Function

mapfoldr(f, rio::RightNeutralOperator, ps)

Equivalent to mapfoldr(f, rio, ps; init = only(right_neutrals(rio))).

See also RightNeutralOperator.

Tip

This also works with foldr(rio, ps).

Examples

julia> @p mapfoldr(not, and, [p, q, r, s])
¬p ∧ (¬q ∧ (¬r ∧ ¬s))

julia> foldr(and, [])
tautology (generic function with 1 method)

Utilities

PAndQ.arity — Function

arity(::LogicalOperator)

Return the arity of the given LogicalOperator.

Examples

julia> arity(tautology)
0

julia> arity(not)
1

julia> arity(and)
2