Printing

PAndQ.TruthTable — Type

TruthTable(ps)

Construct a truth table for the given Propositions and LogicalOperators.

The header is a vector containing vectors of logically equivalent propositions. The sub_header corresponds to the header, but contains each proposition's UnionAll type. The body is a matrix where the rows contain interpretations of each proposition in the given column.

See also tautology and contradiction.

Examples

julia> TruthTable([Tree(⊤)])
┌──────┐
│ ⊤    │
│ Tree │
├──────┤
│ ⊤    │
└──────┘

julia> @atomize TruthTable([¬p])
┌──────────┬─────────┐
│ p        │ ¬p      │
│ Variable │ Literal │
├──────────┼─────────┤
│ ⊤        │ ⊥       │
│ ⊥        │ ⊤       │
└──────────┴─────────┘

julia> @atomize TruthTable([p ∧ ¬p, p ⊻ q, ¬(p ∧ q) ∧ (p ∨ q)])
┌────────┬──────────┬──────────┬───────────────────────────┐
│ p ∧ ¬p │ p        │ q        │ p ⊻ q, ¬(p ∧ q) ∧ (p ∨ q) │
│ Tree   │ Variable │ Variable │ Tree, Tree                │
├────────┼──────────┼──────────┼───────────────────────────┤
│ ⊥      │ ⊤        │ ⊤        │ ⊥                         │
│ ⊥      │ ⊥        │ ⊤        │ ⊤                         │
├────────┼──────────┼──────────┼───────────────────────────┤
│ ⊥      │ ⊤        │ ⊥        │ ⊤                         │
│ ⊥      │ ⊥        │ ⊥        │ ⊥                         │
└────────┴──────────┴──────────┴───────────────────────────┘

Base.show — Function

show(::IO, ::MIME"text/plain", ::Proposition)

Represent the given Proposition as a propositional formula.

Examples

julia> @atomize show(stdout, MIME"text/plain"(), p ⊻ q)
p ⊻ q

julia> @atomize show(stdout, MIME"text/plain"(), Normal(p ⊻ q))
(p ∨ q) ∧ (¬p ∨ ¬q)

show(::IO, ::MIME"text/plain", ::TruthTable)

Examples

julia> @atomize show(stdout, MIME"text/plain"(), TruthTable([p ∧ q]))

show(::IO, ::Proposition)

Represent the given Proposition expanded as valid Julia code.

Examples

julia> @atomize s = sprint(show, p ∧ q)
"Tree(and, Tree(identity, Variable(:p)), Tree(identity, Variable(:q)))"

julia> @eval $(Meta.parse(s))
p ∧ q

PAndQ.formatter — Function

formatter(t::Type{<:Union{PAndQ.NullaryOperator, String, Char, Bool, Int}})

t	formatter(t)(⊤, _, _)	formatter(t)(⊥, _, _)

PAndQ.NullaryOperator	"⊤"	"⊥"

String	"tautology"	"contradiction"

Char	"T"	"F"

Bool	"true"	"false"

Int	"1"	"0"

PrettyTables.pretty_table — Method

pretty_table(
    ::Union{IO, Type{Union{String, Docs.HTML}}} = stdout, ::Union{Proposition, TruthTable};
    formatters = formatter(PAndQ.NullaryOperator), kwargs...
)

See also PrettyTables.pretty_table, Proposition, TruthTable, and formatter.

Examples

julia> pretty_table(@atomize p ∧ q)
┌──────────┬──────────┬───────┐
│ p        │ q        │ p ∧ q │
│ Variable │ Variable │ Tree  │
├──────────┼──────────┼───────┤
│ ⊤        │ ⊤        │ ⊤     │
│ ⊥        │ ⊤        │ ⊥     │
├──────────┼──────────┼───────┤
│ ⊤        │ ⊥        │ ⊥     │
│ ⊥        │ ⊥        │ ⊥     │
└──────────┴──────────┴───────┘

julia> print(pretty_table(Docs.HTML, @atomize p ∧ q).content)
<table>
  <thead>
    <tr class = "header">
      <th style = "text-align: left;">p</th>
      <th style = "text-align: left;">q</th>
      <th style = "text-align: left;">p ∧ q</th>
    </tr>
    <tr class = "subheader headerLastRow">
      <th style = "text-align: left;">Variable</th>
      <th style = "text-align: left;">Variable</th>
      <th style = "text-align: left;">Tree</th>
    </tr>
  </thead>
  <tbody>
    <tr>
      <td style = "text-align: left;">⊤</td>
      <td style = "text-align: left;">⊤</td>
      <td style = "text-align: left;">⊤</td>
    </tr>
    <tr>
      <td style = "text-align: left;">⊥</td>
      <td style = "text-align: left;">⊤</td>
      <td style = "text-align: left;">⊥</td>
    </tr>
    <tr>
      <td style = "text-align: left;">⊤</td>
      <td style = "text-align: left;">⊥</td>
      <td style = "text-align: left;">⊥</td>
    </tr>
    <tr>
      <td style = "text-align: left;">⊥</td>
      <td style = "text-align: left;">⊥</td>
      <td style = "text-align: left;">⊥</td>
    </tr>
  </tbody>
</table>

AbstractTrees.print_tree — Function

print_tree(::IO = stdout, ::Proposition; kwargs...)

Prints a tree diagram of the given Proposition.

Note

Instances of Compound{typeof(identity)} are represented as I.

See also AbstractTrees.print_tree.

julia> @atomize print_tree(p ∧ ¬q ⊻ s)
⊻
├─ ∧
│  ├─ I
│  │  └─ p
│  └─ ¬
│     └─ q
└─ I
   └─ s

julia> @atomize print_tree(Normal(p ∧ ¬q ⊻ s))
∧
├─ ∨
│  ├─ I
│  │  └─ p
│  └─ I
│     └─ s
├─ ∨
│  ├─ ¬
│  │  └─ q
│  └─ I
│     └─ s
└─ ∨
   ├─ ¬
   │  └─ p
   ├─ I
   │  └─ q
   └─ ¬
      └─ s