Propositional Logic

PAQ.Proposition — Type

Proposition

Set of well-formed logical formulae.

Calling an instance of Proposition will return a vector of valid interpretations.

Supertype of Atom, Compound, and Truth. ```

PAQ.Compound — Type

Compound <: Proposition

Compound proposition.

Subtype of Proposition. Supertype of Literal, Tree, and Normal.

PAQ.Atom — Type

Atom <: Proposition
Atom([statement::String])

Atomic proposition.

Info

Constructing an Atom with no argument, an empty string, or an underscore character will set statement = "_". This serves two purposes. Firstly, it is useful as a default proposition when converting Truths to other forms; for example: Tree(⊥) is printed as "_" ∧ ¬"_". Secondly, this ensures that pretty-printing does not produce output such as: ``∧ ¬. It is not idiomatic to use this as a generic proposition; use @atom instead.

Subtype of Proposition.

Examples

julia> p = Atom("p")
Atom:
  "p"

julia> p()
Contingency:
  ["p" => ⊤] => ⊤
  ["p" => ⊥] => ⊥

PAQ.Literal — Type

Literal{
    L <: Union{
        Atom,
        Tuple{Not, Atom}
    }
} <: Compound <: Proposition
Literal(p::L)

An Atom or its negation.

Subtype of Compound and Proposition. See also Not.

Examples

julia> r = ¬p
Literal:
  ¬"p"

julia> r()
Contingency:
  ["p" => ⊤] => ⊥
  ["p" => ⊥] => ⊤

PAQ.Tree — Type

Tree{
    L <: Union{
        Tuple{Not, Compound},
        Tuple{And, Compound, Compound}
    }
} <: Compound <: Proposition
Tree(node::L)

Abstract syntax tree representing a compound proposition.

Note that Not and And are functionally complete operators.

Subtype of Compound and Proposition.

Examples

julia> r = p ∧ ¬p
Tree:
  "p" ∧ ¬"p"

julia> r()
Truth:
  ⊥

julia> (p ∧ q)()
Contingency:
  ["p" => ⊤, "q" => ⊤] => ⊤
  ["p" => ⊤, "q" => ⊥] => ⊥
  ["p" => ⊥, "q" => ⊤] => ⊥
  ["p" => ⊥, "q" => ⊥] => ⊥

PAQ.Normal — Type

Normal{B <: Union{And, Or}} <: Compound <: Proposition
Normal(::Union{typeof(and), typeof(or)}, p::Proposition)

The conjunctive or disjunctive normal form of a proposition.

Constructing an instance with the parameters ([`and`](@ref), p) and ([`or`](@ref), p) correspond to conjunctive and disjunctive normal form, respectively.

Subtype of Compound and Proposition.

Examples

julia> r = Normal(and, p ∧ q)
Normal:
  (¬"p" ∨ "q") ∧ ("p" ∨ ¬"q") ∧ ("p" ∨ "q")

julia> s = Normal(or, ¬p ∨ ¬q)
Normal:
  ("p" ∧ ¬"q") ∨ (¬"p" ∧ "q") ∨ (¬"p" ∧ ¬"q")

julia> t = r ∧ s
Tree:
  (¬"p" ∨ "q") ∧ ("p" ∨ ¬"q") ∧ ("p" ∨ "q") ∧ ("p" ∧ ¬"q") ∨ (¬"p" ∧ "q") ∨ (¬"p" ∧ ¬"q")

julia> t()
Truth:
  ⊥

PAQ.Truth — Type

Truth{V <: Union{Val{:⊥}, Val{:⊤}}} <: Proposition
Truth(::V)

Container for the constants tautology and contradiction. Subtype of Proposition.

PAQ.tautology — Constant

⊤
tautology

A constant which is true in every possible interpretation.

One of two valid instances of Truth, the other instance being contradiction.

⊤ can be typed by \top<tab>.

Examples

julia> ¬⊤
Truth:
  ⊥

julia> tautology()
Truth:
  ⊤

PAQ.contradiction — Constant

⊥
contradiction

A constant which is false in every possible interpretation.

One of two valid instances of Truth, the other instance being tautology.

⊥ can be typed by \bot<tab>.

Examples

julia> ¬⊥
Truth:
  ⊤

julia> contradiction()
Truth:
  ⊥

PAQ.Contingency — Type

Contingency <: Compound

Examples

julia> p()
Contingency:
  ["p" => ⊤] => ⊤
  ["p" => ⊥] => ⊥

julia> (p ∧ q)()
Contingency:
  ["p" => ⊤, "q" => ⊤] => ⊤
  ["p" => ⊤, "q" => ⊥] => ⊥
  ["p" => ⊥, "q" => ⊤] => ⊥
  ["p" => ⊥, "q" => ⊥] => ⊥

PAQ.@atom — Macro

@atom(ps...)

Instantiates atomic propositions.

Examples

julia> @atom p q

julia> p
Atom:
  "p"

julia> q
Atom:
  "q"

PAQ.get_atoms — Function

get_atoms(ps::Proposition...)

Returns a vector of atomic propositions contained in ps.

Warning

Some atoms may optimized out of an expression, such as in p ∧ ⊥.

See also Proposition.

Examples

julia> r = p ∧ q
Tree:
  "p" ∧ "q"

julia> get_atoms(r)
2-element Vector{Atom}:
 "p"
 "q"