Propositions

Tip

Propositions can be converted into different, but logically equivalent forms (see also ==). For example, ⊥ == Tree(⊥) == Clause(⊥) == Normal(⊥). However, not all forms are Expressively complete. Otherwise, the conversion may throw an exception. For example, there is no way to represent Literal(⊥).

Proposition
├─ Atom
│  ├─ Constant
│  └─ Variable
└─ Compound
   ├─ Clause
   ├─ Expressive
   │  ├─ Normal
   │  └─ Tree
   └─ Literal

Abstract Types

PAndQ.Proposition — Type

Proposition

The set of well-formed logical formulae.

Supertype of Atom and Compound.

PAndQ.Atom — Type

Atom <: Proposition

A proposition with no deeper propositional structure.

Tip

Use @atomize and @variables for syntax sugar to instantiate atoms inline and define variables, respectively.

Subtype of Proposition. Supertype of Constant and Variable.

PAndQ.Compound — Type

Compound{LO} <: Proposition

A proposition composed from connecting Atomic propositions with LogicalOperators.

Subtype of Proposition. Supertype of Literal, Clause, and Expressive.

PAndQ.Expressive — Type

Expressive{LO} <: Compound{LO}

A proposition that is expressively complete.

Subtype of Compound. Supertype of Tree and Normal.

Concrete Types

PAndQ.Constant — Type

Constant{T} <: Atom
Constant(::T)

An atomic sentence.

Tip

Define pretty-printing for an instance of Constant{T} by overloading show(io::IO, ::MIME"text/plain", p::Constant{T}).

Subtype of Atom.

Examples

julia> Constant(1)
$(1)

julia> Constant("Logic is fun")
$("Logic is fun")

PAndQ.Variable — Type

Variable <: Atom

An atomic formula.

Subtype of Atom.

Examples

julia> Variable(:p)
p

julia> Variable(:q)
q

PAndQ.Literal — Type

Literal{UO <: UnaryOperator, A <: Atom} <: Compound{UO}
Literal(::UO, ::A)
Literal(::LiteralProposition)

A proposition represented by an atomic formula or its negation.

Subtype of Compound. See also UnaryOperator, Atom, and LiteralProposition.

Examples

julia> @atomize ¬p
¬p

PAndQ.Tree — Type

Tree{LO <: LogicalOperator, AT <: Union{Atom, Tree}} <: Expressive{LO}
Tree(::NullaryOperator, ::Atom)
Tree(::LogicalOperator, ::Tree...)
Tree(::Proposition)

A proposition represented by an abstract syntax tree.

Subtype of Expressive. See also LogicalOperator.

Examples

julia> @atomize p ⊻ q
p ⊻ q

julia> @atomize ¬p → q
¬p → q

PAndQ.Clause — Type

Clause{AO <: AndOr, L <: Literal} <: Compound{AO}
Clause(::AO, ps = Literal[])
Clause(::AO, p::Proposition)
Clause(::Union{NullaryOperator, LiteralProposition})

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty clause is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also AndOr, Literal, NullaryOperator, and LiteralProposition.

Examples

julia> Clause(and)
⊤

julia> @atomize Clause(p)
p

julia> @atomize Clause(or, [¬p, q])
¬p ∨ q

PAndQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Expressive{AO}
Normal(::typeof(and), ps = Clause{typeof(or)}[])
Normal(::typeof(or), ps = Clause{typeof(and)}[])
Normal(::AO, ::Proposition)
Normal(::Union{NullaryOperator, Proposition})

A proposition represented in conjunctive or disjunctive normal form.

Info

An empty normal form is logically equivalent to the neutral element of it's binary operator.

Subtype of Expressive. See also AndOr, Clause, NullaryOperator, and Proposition.

Examples

julia> @atomize s = Normal(and, [Clause(or, [p, q]), Clause(or, ¬r)])
(p ∨ q) ∧ (¬r)

julia> ¬s
(¬p ∧ ¬q) ∨ (r)

Macros

PAndQ.@atomize — Macro

@atomize(expression)

Instantiate undefined variables and interpolated values inline as Atoms.

Warning

This macro attempts to ignore symbols that are being assigned a value. For example, @atomize f(; x = p) = x ∧ q should be equivalent to @atomize f(; x = Atom(:p)) = x ∧ Atom(:q). However, this feature is in-progress and only works in some cases. The implementation is cautious to skip the parts of the expression that it cannot yet handle.

Examples

julia> @atomize x = p ∧ q
p ∧ q

julia> @atomize x → r
(p ∧ q) → r

julia> @atomize $1 ∧ $(1 + 1)
$(1) ∧ $(2)

PAndQ.@variables — Macro

@variables(xs...)

Define variables and return a vector containing them.

Expressions of the form symbol and symbol = value are defined as symbol = @atomize symbol and symbol = @atomize value, respectively.

See also @atomize.

Examples

julia> @variables p q = ¬$1
2-element Vector{Proposition}:
 p
 ¬$(1)

julia> p
p

julia> q
¬$(1)

Utilities

PAndQ.atoms — Function

atoms(p, T = Atom)

Return an iterator of each Atom of type T contained in p.

See also Constant and Variable.

Examples

julia> @atomize collect(atoms(p ∧ q))
2-element Vector{Variable}:
 p
 q

julia> @atomize collect(atoms(p ∧ q ∨ $1 ∧ $2, Constant))
2-element Vector{Constant{Int64}}:
 $(1)
 $(2)

PAndQ.operators — Function

operators(p)

Return an iterator of each LogicalOperator contained in p.

Examples

julia> @atomize collect(operators(¬p))
1-element Vector{typeof(not)}:
 not (generic function with 19 methods)

julia> @atomize collect(operators(¬p ∧ q))
3-element Vector{Function}:
 and (generic function with 18 methods)
 not (generic function with 19 methods)
 identity (generic function with 1 method)

Base.map — Function

map(f, ::Proposition)

Apply f to each Atom in the Proposition.

Examples

julia> @atomize map(Tree ∘ ¬, p ∧ q)
¬p ∧ ¬q

julia> @atomize map(p ∧ q) do atom
           println(atom)
           atom
       end
Variable(:p)
Variable(:q)
p ∧ q