Semantics

Warning

The model checking algorithm currently has a naive implementation with a time complexity of O(2 ^ length(unique(atoms(p)))).

Truths

PAndQ.valuations — Function

valuation(atoms)
valuations(::Union{NullaryOperator, Proposition})

Return an iterator of every possible valuation.

See also Nullary Operators and Proposition.

Examples

julia> collect(valuations(⊤))
0-dimensional Array{Vector{Union{}}, 0}:
[]

julia> @atomize collect(valuations(p))
2-element Vector{Vector{Pair{PAndQ.Variable, Bool}}}:
 [PAndQ.Variable(:p) => 1]
 [PAndQ.Variable(:p) => 0]

julia> @atomize collect(valuations(p ∧ q))
2×2 Matrix{Vector{Pair{PAndQ.Variable, Bool}}}:
 [Variable(:p)=>1, Variable(:q)=>1]  [Variable(:p)=>1, Variable(:q)=>0]
 [Variable(:p)=>0, Variable(:q)=>1]  [Variable(:p)=>0, Variable(:q)=>0]

PAndQ.interpret — Function

interpret(valuation, ::Union{NullaryOperator, Proposition})

Substitute each Atom in the given Proposition with values from the valuation.

The valuation can be a function with the signature valuation(::Atom)::Union{Bool, NullaryOperator, Proposition}, a Dict, or an iterable that can construct a Dict. No substitution is performed if an Atom from the Proposition is not one of the dictionary's keys.

See also Nullary Operators.

Examples

julia> @atomize interpret(p -> true, ¬p)
false

julia> @atomize interpret(p => ⊤, p ∧ q)
⊤ ∧ q

PAndQ.interpretations — Function

interpretations(valuations, p)
interpretations(p)

Return an iterator of truth values given by interpreting p with each valuation in valuations.

Examples

julia> collect(interpretations(⊤))
0-dimensional Array{Bool, 0}:
1

julia> @atomize collect(interpretations(p))
2-element Vector{Bool}:
 1
 0

julia> @atomize collect(interpretations(p ∧ q))
2×2 Matrix{Bool}:
 1  0
 0  0

PAndQ.solve — Function

solve(valuations, p)
solve(p)

Return a vector containing all valuations such that interpret(valuation, p) == ⊤.

See also interpret and tautology.

Examples

julia> collect(solve(⊤))
1-element Vector{Vector{Union{}}}:
 []

julia> @atomize collect(solve(p))
1-element Vector{Vector{Pair{PAndQ.Variable, Bool}}}:
 [PAndQ.Variable(:p) => 1]

julia> @atomize collect(solve(p ∧ q))
1-element Vector{Vector{Pair{PAndQ.Variable, Bool}}}:
 [PAndQ.Variable(:p) => 1, PAndQ.Variable(:q) => 1]

Predicates

PAndQ.is_commutative — Function

is_commutative(::BinaryOperator)

See also Binary Operators.

PAndQ.is_associative — Function

is_associative(::BinaryOperator)

See also Binary Operators.

Base.:== — Function

==(::Union{Bool, NullaryOperator, Proposition}, ::Union{Bool, NullaryOperator, Proposition})
p == q

Returns a boolean indicating whether p and q are logically equivalent.

Constants are equivalent if and only if their values are equivalent.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it.

See also Nullary Operators and Proposition.

Examples

julia> @atomize p == ¬p
false

julia> @atomize ¬(p ⊻ q) == (p → q) ∧ (p ← q)
true

julia> @atomize ¬(p ⊻ q) === (p → q) ∧ (p ← q)
false

PAndQ.is_tautology — Function

is_tautology(p)

Returns a boolean on whether p is a tautology.

Examples

julia> is_tautology(⊤)
true

julia> @atomize is_tautology(p)
false

julia> @atomize is_tautology(¬(p ∧ ¬p))
true

PAndQ.is_contradiction — Function

is_contradiction(p)

Returns a boolean on whether p is a contradiction.

Examples

julia> is_contradiction(⊥)
true

julia> @atomize is_contradiction(p)
false

julia> @atomize is_contradiction(p ∧ ¬p)
true

PAndQ.is_truth — Function

is_truth(p)

Returns a boolean on whether p is a nullary operator.

See also Proposition.

Examples

julia> is_truth(⊤)
true

julia> @atomize is_truth(p ∧ ¬p)
true

julia> @atomize is_truth(p)
false

julia> @atomize is_truth(p ∧ q)
false

PAndQ.is_contingency — Function

is_contingency(p)

Returns a boolean on whether p is a contingency (not logically equivalent to a nullary operator).

See also Proposition.

Examples

julia> is_contingency(⊤)
false

julia> @atomize is_contingency(p ∧ ¬p)
false

julia> @atomize is_contingency(p)
true

julia> @atomize is_contingency(p ∧ q)
true

PAndQ.is_satisfiable — Function

is_satisfiable(p)

Returns a boolean on whether p is satisfiable (not a contradiction).

See also Proposition.

Examples

julia> is_satisfiable(⊤)
true

julia> @atomize is_satisfiable(p ∧ ¬p)
false

julia> @atomize is_satisfiable(p)
true

julia> @atomize is_satisfiable(p ∧ q)
true

PAndQ.is_falsifiable — Function

is_falsifiable(p)

Returns a boolean on whether p is falsifiable (not a tautology).

See also Proposition.

Examples

julia> is_falsifiable(⊥)
true

julia> @atomize is_falsifiable(p ∨ ¬p)
false

julia> @atomize is_falsifiable(p)
true

julia> @atomize is_falsifiable(p ∧ q)
true

Properties

PAndQ.dual — Function

dual(::Operator)

Returns the operator that is the dual of the given operator.

Examples

julia> dual(and)
or (generic function with 17 methods)

julia> @atomize and(p, q) == not(dual(and)(not(p), not(q)))
true

julia> dual(imply)
not_converse_imply (generic function with 3 methods)

julia> @atomize imply(p, q) == not(dual(imply)(not(p), not(q)))
true

PAndQ.converse — Function

converse(::Operator)

Returns the operator that is the converse of the given operator.

Examples

julia> converse(and)
and (generic function with 17 methods)

julia> @atomize and(p, q) == converse(and)(q, p)
true

julia> converse(imply)
converse_imply (generic function with 3 methods)

julia> @atomize imply(p, q) == converse(imply)(q, p)
true

PAndQ.left_neutrals — Function

left_neutrals(::Operator)

Return a Set of the Nullary Operators that are left neutral elements of the given operator.

Examples

julia> left_neutrals(or)
Set{typeof(contradiction)} with 1 element:
  PAndQ.contradiction

julia> left_neutrals(imply)
Set{typeof(tautology)} with 1 element:
  PAndQ.tautology

julia> left_neutrals(nor)
Set{Union{typeof(contradiction), typeof(tautology)}}()

PAndQ.right_neutrals — Function

right_neutrals(::Operator)

Return a Set of the Nullary Operators that are right neutral elements of the given operator.

Examples

julia> right_neutrals(or)
Set{typeof(contradiction)} with 1 element:
  PAndQ.contradiction

julia> right_neutrals(converse_imply)
Set{typeof(tautology)} with 1 element:
  PAndQ.tautology

julia> left_neutrals(nor)
Set{Union{typeof(contradiction), typeof(tautology)}}()

Utilities

Core.Bool — Method

Bool(::NullaryOperator)

See also Nullary Operators.

Base.convert — Method

convert(::Type{Bool}, ::NullaryOperator)

See also Nullary Operators.