Utility

PAQ.@atoms — Macro

@atoms(ps...)

Instantiate and define Atoms with symbols and return a vector containing them.

Info

Atoms are defined in the global scope as constants.

Examples

julia> @atoms p q
2-element Vector{Atom{Symbol}}:
 p
 q

julia> p
Atom:
 p

julia> q
Atom:
 q

PAQ.@p — Macro

@p(x)

Returns a propositions by instantiating all strings and undefined variables as Atoms, and then evaluating the expression.

Examples

julia> x = @p p
Atom:
 p

julia> @p x ∧ q → "r"
Tree:
 (p ∧ q) → "r"

PAQ.@p_str — Macro

@p_str(x)

Examples

julia> p = @p_str("x")
Atom:
 x

julia> p"\"p\" ∧ p, Clause(and)"
("p" ∧ x, ⊤)

PAQ.get_atoms — Function

get_atoms(::Proposition...)

Returns a vector of unique Atoms contained in the given Proposition(s).

Warning

Some atoms may optimized out of an expression, such as in p ∧ ⊥ == ⊥.

Examples

julia> @p get_atoms(p ∧ q)
2-element Vector{Atom{Symbol}}:
 p
 q