Types

PAQ.Language — Type

Language

Set of well-formed logical formulae.

Supertype of Primitive, Compound, and Truth.

PAQ.Compound — Type

Compound <: Language

Compound proposition.

Subtype of Language. Supertype of Propositional.

PAQ.Primitive — Type

Primitive <: Language
Primitive([statement::String = ""])

Primitive proposition.

Subtype of Language. See also Compound.

Examples

julia> p = Primitive("Logic is fun")
Primitive("Logic is fun")

julia> ¬p
Propositional(
  Not(), Propositional(
    Primitive("Logic is fun")
  )
)

PAQ.Operator — Type

Operator

Set of functions that operate on a logical Language.

Supertype of Boolean.

PAQ.Boolean — Type

Boolean <: Operator

Set of functionally complete logical connectives.

Subtype of Operator. Supertype of Not and And. See also Boolean Operators.

PAQ.Not — Type

Not <: Boolean <: Operator

Singleton type representing logical negation.

Subtype of Boolean and Operator. See also And and Language.

Examples

julia> @truth_table PAQ.Not()(p)
┌───────────┬────────────────┐
│         p │ (PAQ.Not())(p) │
│ Primitive │  Propositional │
├───────────┼────────────────┤
│         ⊤ │              ⊥ │
│         ⊥ │              ⊤ │
└───────────┴────────────────┘

PAQ.And — Type

And <: Boolean <: Operator

Singleton type representing logical conjunction.

Subtype of Boolean and Operator. See also Not and Language.

Examples

julia> @truth_table PAQ.And()(p, q)
┌───────────┬───────────┬───────────────────┐
│         p │         q │ (PAQ.And())(p, q) │
│ Primitive │ Primitive │     Propositional │
├───────────┼───────────┼───────────────────┤
│         ⊤ │         ⊤ │                 ⊤ │
│         ⊤ │         ⊥ │                 ⊥ │
├───────────┼───────────┼───────────────────┤
│         ⊥ │         ⊤ │                 ⊥ │
│         ⊥ │         ⊥ │                 ⊥ │
└───────────┴───────────┴───────────────────┘

PAQ.Propositional — Type

Propositional{
    L <: Union{
        Primitive,
        Tuple{Not, Compound},
        Tuple{And, Compound, Compound}
    }
} <: Compound <: Language
Propositional(ϕ::L)

Abstract syntax tree representing a compound proposition.

Subtype of Compound and Language. See also Primitive, Not, and And.

Examples

julia> p = Propositional(Primitive("p"))
Propositional(
  Primitive("p")
)

julia> ¬p
Propositional(
  Not(), Propositional(
    Primitive("p")
  )
)

PAQ.Truth — Type

Truth{V <: Union{Val{:⊥}, Val{:⊤}}} <: Language
Truth(::V)

Container for the constants tautology and contradiction. Subtype of Language.

PAQ.tautology — Constant

⊤
tautology

A constant which is true in every possible interpretation.

One of two valid instances of Truth, the other instance being contradiction.

⊤ can be typed by \top<tab>.

Examples

julia> ¬⊤
⊥

julia> tautology()
⊤

PAQ.contradiction — Constant

⊥
contradiction

A constant which is false in every possible interpretation.

One of two valid instances of Truth, the other instance being tautology.

⊥ can be typed by \bot<tab>.

Examples

julia> ¬⊥
⊤

julia> contradiction()
⊥