Utility

PAQ.get_primitives — Function

get_primitives(ps::Language...)

Returns a vector of Primitive propositions contained in p.

Note that some primitives may optimized out of an expression, such as in p ∧ ⊥.

See also Language.

Examples

julia> get_primitives(p)
1-element Vector{Primitive}:
 Primitive("p")

julia> get_primitives(p ∧ q, r)
3-element Vector{Primitive}:
 Primitive("p")
 Primitive("q")
 Primitive("r")

PAQ.@primitive — Macro

@primitive(ps...)

Instantiates Primitive propositions.

Examples

julia> @primitive p q

julia> p
Primitive("p")

julia> q
Primitive("q")

PAQ.@truth_table — Macro

@truth_table p
@truth_table(ps::Language...)

Print a truth table for the given propositions.

See also Language.

Examples

julia> @truth_table p∧q p→q
┌────────────────┬────────────────┬───────────────┬───────────────┐
│              p │              q │         p ∧ q │         p → q │
│ Primitive("p") │ Primitive("q") │ Propositional │ Propositional │
├────────────────┼────────────────┼───────────────┼───────────────┤
│              ⊤ │              ⊤ │             ⊤ │             ⊤ │
│              ⊤ │              ⊥ │             ⊥ │             ⊥ │
├────────────────┼────────────────┼───────────────┼───────────────┤
│              ⊥ │              ⊤ │             ⊥ │             ⊤ │
│              ⊥ │              ⊥ │             ⊥ │             ⊤ │
└────────────────┴────────────────┴───────────────┴───────────────┘

PAQ.interpret — Function

interpret(valuation, ϕ)

Given a valuation function that maps from the Primitive propositions in ϕ to their respective Truth values, assign a truth value to ϕ.

See also Language.

julia> mapping = Dict(p => ⊥, q => ⊤);

julia> valuation = r -> mapping[r];

julia> interpret(valuation, p ∧ q)
⊥

julia> interpret(valuation, p → q)
⊤

Base.:== — Function

p == q
==(p::Language, q::Language)
isequal(p::Language, q::Language)

Returns a boolean indicating whether p and q are logically equivalent.

See also Language.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it.

Examples

julia> p == ¬p
false

julia> julia> p ∨ q == ¬(¬q ∧ ¬p)
true

julia> isequal((p → q) ∧ (p ← q), ¬(p ⊻ q))
true