Internals

Operators

PAQ.NullaryOperator — Type

NullaryOperator

The union types of BooleanOperators that take zero arguments.

PAQ.UnaryOperator — Type

UnaryOperator

The union types of BooleanOperators that take one argument.

PAQ.BinaryOperator — Type

BinaryOperator

The union types of BooleanOperators that take two arguments.

PAQ.NaryOperator — Type

NaryOperator

PAQ.BooleanOperator — Type

BooleanOperator

The union types of boolean operators.

PAQ.CommutativeOperator — Type

CommutativeOperator

The union types of BooleanOperators with the commutative property.

PAQ.AssociativeOperator — Type

AssociativeOperator

The union types of BooleanOperators with the associative property.

PAQ.LeftIdentityOperator — Type

LeftIdentityOperator

The union types of BooleanOperators that have a left identity.

PAQ.RightIdentityOperator — Type

RightIdentityOperator

The union types of BooleanOperators that have a right identity.

PAQ.AndOr — Type

AndOr

The union types of and and or.

Propositions

PAQ.AtomicProposition — Type

AtomicProposition

A Proposition that is known by its type to be equivalent to an Atom.

PAQ.LiteralProposition — Type

LiteralProposition

A Proposition that is known by its type to be equivalent to a Literal.

PAQ.NonExpressive — Type

NonExpressive

A Proposition that is not an Expressive.

Printing

Base.show — Function

show

Base.print — Function

print

Utility

Base.mapfoldl — Function

mapfoldl

Base.mapfoldr — Function

mapfoldr

Base.mapreduce — Function

mapreduce

Semantics

Base.convert — Function

convert