Internals

Operators

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of LogicalOperators that take zero arguments.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of LogicalOperators that take one argument.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of LogicalOperators that take two arguments.

PAndQ.LogicalOperator — Type

LogicalOperator

The Union of logical operators.

PAndQ.CommutativeOperator — Type

CommutativeOperator

The Union of LogicalOperators with the commutative property.

PAndQ.AssociativeOperator — Type

AssociativeOperator

The Union of LogicalOperators with the associative property.

PAndQ.LeftNeutralOperator — Type

LeftNeutralOperator

The Union of LogicalOperators that have one or more left_neutrals.

PAndQ.RightNeutralOperator — Type

RightNeutralOperator

The Union of LogicalOperators that have one or more right_neutrals.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

PAndQ.AtomicProposition — Type

AtomicProposition

A Proposition that is known by its type to be logically equivalent to an Atom.

PAndQ.LiteralProposition — Type

LiteralProposition

A Proposition that is known by its type to be logically equivalent to a Literal.

Printing

Base.show — Function

show

Utility

Base.mapfoldl — Function

mapfoldl(f, lio::LeftNeutralOperator, ps)

Equivalent to mapfoldl(f, lio, ps; init = only(left_neutrals(lio)))

Tip

This also works with foldl(lio, ps).

Examples

julia> @p mapfoldl(not, and, [p, q, r, s])
((¬p ∧ ¬q) ∧ ¬r) ∧ ¬s

julia> foldl(and, [])
tautology (generic function with 1 method)

Base.mapfoldr — Function

mapfoldr(f, rio::RightNeutralOperator, ps)

Equivalent to mapfoldr(f, rio, ps; init = only(right_neutrals(rio)))

Tip

This also works with foldr(rio, ps).

Examples

julia> @p mapfoldr(not, and, [p, q, r, s])
¬p ∧ (¬q ∧ (¬r ∧ ¬s))

julia> foldr(and, [])
tautology (generic function with 1 method)

Semantics

Base.convert — Function

convert