Internals

Everything that is not exported is considered internal.

PAndQ.union_typeof — Function

union_typeof

Operators

Union Types

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of LogicalOperators that take zero arguments.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of LogicalOperators that take one argument.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of LogicalOperators that take two arguments.

PAndQ.LogicalOperator — Type

LogicalOperator

The Union of logical operators.

PAndQ.CommutativeOperator — Type

CommutativeOperator

The Union of LogicalOperators with the commutative property.

PAndQ.AssociativeOperator — Type

AssociativeOperator

The Union of LogicalOperators with the associative property.

PAndQ.LeftNeutralOperator — Type

LeftNeutralOperator

The Union of LogicalOperators that have one or more left_neutrals.

PAndQ.RightNeutralOperator — Type

RightNeutralOperator

The Union of LogicalOperators that have one or more right_neutrals.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

Union Types

PAndQ.AtomicProposition — Type

AtomicProposition

A Proposition that is known by its type to be logically equivalent to an Atom.

PAndQ.LiteralProposition — Type

LiteralProposition

A Proposition that is known by its type to be logically equivalent to a Literal.

AbstractTrees.jl

AbstractTrees.children — Function

children(::Proposition)

Return an iterator over the leaf nodes of the given Proposition.

Warning

Nodes that apply the identity function are skipped.

Examples

julia> @p PAndQ.children(Literal(identity, p))
()

julia> @p PAndQ.children(Literal(not, p))
(Atom(:p),)

julia> @p PAndQ.children(p ⊻ q)
2-element Vector{Tree{typeof(identity), Atom{Symbol}}}:
 p
 q

julia> @p PAndQ.children(Normal(and, p ⊻ q))
2-element Vector{Clause{typeof(or)}}:
 p ∨ q
 ¬p ∨ ¬q

AbstractTrees.printnode — Function

printnode(::IO, ::Proposition)

Warning

Nodes that apply the identity function are skipped.

Examples

julia> @p PAndQ.printnode(stdout, Literal(identity, p))
p
julia> @p PAndQ.printnode(stdout, Literal(not, p))
¬
julia> @p PAndQ.printnode(stdout, p ⊻ q)
⊻
julia> @p PAndQ.printnode(stdout, Normal(and, p ⊻ q))
∧

Utility

PAndQ.union_all_type — Function

union_all_type(::Proposition)

Return the UnionAll type of a Proposition.

Examples

julia> @p PAndQ.union_all_type(p)
Atom

julia> @p PAndQ.union_all_type(p ∧ q)
Tree

PAndQ.only_field — Function

only_field(::Proposition)

Return the only field of a Proposition.

Examples

julia> @p PAndQ.only_field(p)
:p

julia> @p PAndQ.only_field(p ∧ q)
2-element Vector{Tree{typeof(identity), Atom{Symbol}}}:
 p
 q

PAndQ.operator_to_proposition — Function

operator_to_proposition(::Union{LogicalOperator, Proposition})

Construct a Proposition with the given operator.

Examples

julia> PAndQ.operator_to_proposition(⊤)
⊤

julia> PAndQ.operator_to_proposition(¬)
¬_

julia> PAndQ.operator_to_proposition(∧)
_ ∧ __

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as an Atom if it is undefined. If x is an expression, traverse it with recursive calls to atomize (ignoring variable assignment and keyword arguments). Otherise, return x.

Printing

PAndQ.operator_to_symbol — Function

operator_to_symbol(::LogicalOperator)

Return the Unicode character that is an alias for the given LogicalOperator.

Examples

julia> PAndQ.operator_to_symbol(⊤)
"⊤"

julia> PAndQ.operator_to_symbol(¬)
"¬"

julia> PAndQ.operator_to_symbol(∧)
"∧"

PAndQ.merge_string — Function

merge_string(cell)

PAndQ.letter — Function

letter(::NullaryOperator)

Examples

julia> PAndQ.letter(tautology)
"T"

julia> PAndQ.letter(contradiction)
"F"

PAndQ.format_latex — Function

format_latex(x)

PAndQ.format_head — Function

format_head(format, cell)

PAndQ.format_body — Function

format_body(format, cell)

PAndQ._newline — Function

_newline(::Bool)

PAndQ.print_string — Function

print_string

PAndQ.parenthesize — Function

parenthesize(::IO, x)

Semantics

PAndQ.CallableObjectDocumentation — Type

CallableObjectDocumentation

A dummy type to attach a docstring to a callable object.

See also this Documenter.jl issue

PAndQ.neutral_operator — Function

neutral_operator(::NullaryOperator)

Return a subtype of AndOr that is the neutral element of the given NullaryOperator.

See also left_neutrals and right_neutrals.

Examples

julia> PAndQ.neutral_operator(tautology)
and (generic function with 23 methods)

julia> PAndQ.neutral_operator(contradiction)
or (generic function with 19 methods)

PAndQ.eval_doubles — Function

eval_doubles(f, doubles)