Internals

Everything that is not exported is considered internal.

PAndQ.union_typeof — Function

union_typeof(xs)

Operators

PAndQ.FoldDirection — Type

FoldDirection

A trait to indicate which direction to fold an operator.

Supertype of Left and Right.

PAndQ.Left — Type

Left <: FoldDirection

A trait to indicate that an operator should fold left.

Subtype of FoldDirection.

PAndQ.Right — Type

Right <: FoldDirection

A trait to indicate that an operator should fold right.

Subtype of FoldDirection.

PAndQ.fold_direction — Function

fold_direction(operator)

Return the FoldDirection of the given operator.

If the operator has no FoldDirection, return nothing.

julia> PAndQ.fold_direction(→)
PAndQ.Left()

julia> PAndQ.fold_direction(←)
PAndQ.Right()

julia> PAndQ.fold_direction(⊼)

Union Types

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of Nullary Operators.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of Unary Operators.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of Binary Operators.

PAndQ.NaryOperator — Type

NaryOperator

The Union of Nary Operators.

PAndQ.Operator — Type

Operator

The Union of Operators.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

PAndQ.value_exception — Constant

value_exception

PAndQ.simplify_clause — Function

simplify_clause(p)

Types

Proposition
├─ Atom
│  ├─ Constant
│  └─ Variable
└─ Compound
   ├─ Clause
   ├─ Normal
   └─ Tree

Abstract

PAndQ.Proposition — Type

Proposition

A proposition.

Supertype of Atom and Compound.

PAndQ.Atom — Type

Atom <: Proposition

A proposition with no deeper propositional structure.

Subtype of Proposition. Supertype of Constant and Variable.

PAndQ.Compound — Type

Compound <: Proposition

A proposition composed from connecting Atoms with one or more Operators.

Subtype of Proposition. Supertype of Tree, Clause, and Normal.

Concrete

PAndQ.Constant — Type

Constant{T} <: Atom
Constant(::T)

An atomic sentence.

Subtype of Atom.

Examples

julia> PAndQ.Constant(1)
$(1)

julia> PAndQ.Constant("Logic is fun")
$("Logic is fun")

PAndQ.Variable — Type

Variable <: Atom

An atomic formula.

Subtype of Atom.

Examples

julia> PAndQ.Variable(:p)
p

julia> PAndQ.Variable(:q)
q

PAndQ.Tree — Type

Tree{O <: Operator, AT <: Union{Atom, Tree}, N} <: Compound
Tree(::NullaryOperator, ::Atom)
Tree(::Operator, ::Tree...)
Tree(::Proposition)

A Proposition represented by an abstract syntax tree.

Subtype of Compound. See also Atom, NullaryOperator, and Operator.

Examples

julia> PAndQ.Tree(⊤)
⊤

julia> @atomize PAndQ.Tree(¬, p)
¬p

julia> @atomize PAndQ.Tree(and, PAndQ.Tree(p), PAndQ.Tree(q))
p ∧ q

PAndQ.Literal — Type

Literal <: Tree{<:Operator, <:Atom, 1}

A proposition represented by an atomic formula or its negation.

See also Operator, Atom, and Tree.

Examples

julia> @atomize PAndQ.Literal(𝒾, p)
p

julia> @atomize PAndQ.Literal(¬, p)
¬p

PAndQ.Clause — Type

Clause{AO <: AndOr, A <: AbstractVector{<:Atom}, L <: AbstractVector{Int}} <: Compound
Clause(::AO, ::A, ::L)

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty Clause is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also Atom, Literal, AndOr, and NullaryOperator.

Examples

julia> PAndQ.Clause(∧, PAndQ.Atom[], Int[])
⊤

julia> @atomize PAndQ.Clause(∧, [p], [1])
p

julia> @atomize PAndQ.Clause(∨, [p, q], [1, -2])
p ∨ ¬q

PAndQ.Normal — Type

Normal{AO <: AndOr, A <: AbstractVector{<:Atom}, C <: AbstractVector{<:AbstractVector{Int}}} <: Compound
Normal(::AO, ::A, ::C)

A Proposition represented in conjunctive or disjunctive normal form.

Info

An empty Normal is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also Clause, AndOr, and NullaryOperator.

Examples

julia> PAndQ.Normal(∧, PAndQ.Atom[], Vector{Int}[])
⊤

julia> @atomize PAndQ.Normal(∧, [p, q], [[1, 2], [-1, -2]])
(p ∨ q) ∧ (¬p ∨ ¬q)

AbstractTrees.jl

AbstractTrees.children — Function

children(::Proposition)

Return an iterator over the child nodes of the given Proposition.

Examples

julia> @atomize PAndQ.children(p)
()

julia> @atomize PAndQ.children(¬p)
(PAndQ.Variable(:p),)

julia> @atomize PAndQ.children(p ∧ q)
(PAndQ.Variable(:p), PAndQ.Variable(:q))

AbstractTrees.nodevalue — Function

nodevalue(::Union{Tree{O}, Clause{O}, Normal{O}}) where O

Return O.instance.

See also Compound.

Examples

julia> @atomize PAndQ.nodevalue(¬p)
not (generic function with 6 methods)

julia> @atomize PAndQ.nodevalue(p ∧ q)
and (generic function with 11 methods)

AbstractTrees.printnode — Function

printnode(::IO, ::Union{NullaryOperator, Proposition}; kwargs...)

See also Proposition.

Examples

julia> @atomize PAndQ.printnode(stdout, p)
p
julia> @atomize PAndQ.printnode(stdout, ¬p)
¬
julia> @atomize PAndQ.printnode(stdout, p ∧ q)
∧

AbstractTrees.NodeType — Type

NodeType(::Type{<:Atom})

See also Atom.

AbstractTrees.nodetype — Function

nodetype(::Type{<:Atom})

See also Atom.

Utility

PAndQ.child — Function

child(x)

Equivalent to only ∘ children

See also children.

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as a Variable if it is undefined in the caller's scope. If isexpr(x, :$), return an expression that instantiates it as a Constant. If x is a different expression, traverse it with recursive calls to atomize. Otherise, return x.

PAndQ.negated_normal — Function

negated_normal(p)

PAndQ.distribute — Function

distribute(p)

PAndQ.flatten! — Function

flatten!(p, clauses)

Printing

PAndQ.symbol_of — Function

symbol_of(::Operator)

Return the Unicode symbol of the given Operator.

Examples

julia> PAndQ.symbol_of(⊤)
:⊤

julia> PAndQ.symbol_of(¬)
:¬

julia> PAndQ.symbol_of(∧)
:∧

PAndQ.parenthesize — Function

parenthesize(io, p)

PAndQ.print_node — Function

print_node(io, p)

PAndQ.show_atom — Function

show_atom(::IO, ::Atom)

See also Atom.

PAndQ.base_type — Function

base_type(::Union{<:Type{<:Proposition}}, Proposition)

Semantics

Base.IteratorSize — Type

IteratorSize(::Type{PicoSolIterator})

Patch a bug in PicoSAT.jl

Warning

This method commits type piracy. Remove it after the PR to fix the bug has been released.

PAndQ.neutral_operator — Function

neutral_operator(::NullaryOperator)

Return one of AndOr that is the neutral element of the given NullaryOperator.

Examples

julia> PAndQ.neutral_operator(⊤)
and (generic function with 11 methods)

julia> PAndQ.neutral_operator(⊥)
or (generic function with 9 methods)

PAndQ.eval_doubles — Function

eval_doubles(f, doubles)

PAndQ.combine — Function

combine(p, q)

PAndQ.negated_normal_template — Function

negated_normal_template(left, right)

Base.convert — Method

convert(::Type{<:Proposition}, p)

See also Proposition.