Internals

Everything that is not exported is considered internal.

PAndQ.union_typeof — Function

union_typeof(xs)

Operators

Union Types

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of Nullary Operators.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of Unary Operators.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of Binary Operators.

PAndQ.NaryOperator — Type

NaryOperator

The Union of Nary Operators.

PAndQ.Operator — Type

Operator

The Union of Operators.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

Types

Proposition
├─ Atom
│  ├─ Constant
│  └─ Variable
└─ Compound
   ├─ Clause
   ├─ Normal
   └─ Tree

Abstract

PAndQ.Proposition — Type

Proposition

A proposition.

Supertype of Atom and Compound.

PAndQ.Atom — Type

Atom <: Proposition

A proposition with no deeper propositional structure.

Subtype of Proposition. Supertype of Constant and Variable.

PAndQ.Compound — Type

Compound <: Proposition

A proposition composed from connecting Atoms with one or more Operators.

Subtype of Proposition. Supertype of Tree, Clause, and Normal.

Concrete

PAndQ.Constant — Type

Constant{T} <: Atom
Constant(::T)

An atomic sentence.

Subtype of Atom.

Examples

julia> PAndQ.Constant(1)
$(1)

julia> PAndQ.Constant("Logic is fun")
$("Logic is fun")

PAndQ.Variable — Type

Variable <: Atom

An atomic formula.

Subtype of Atom.

Examples

julia> PAndQ.Variable(:p)
p

julia> PAndQ.Variable(:q)
q

PAndQ.Tree — Type

Tree{O <: Operator, AT <: Union{Atom, Tree}, N} <: Compound
Tree(::NullaryOperator, ::Atom)
Tree(::Operator, ::Tree...)
Tree(::Proposition)

A Proposition represented by an abstract syntax tree.

Subtype of Compound. See also Atom, NullaryOperator, and Operator.

Examples

julia> PAndQ.Tree(⊤)
⊤

julia> @atomize PAndQ.Tree(¬, p)
¬p

julia> @atomize PAndQ.Tree(and, PAndQ.Tree(p), PAndQ.Tree(q))
p ∧ q

PAndQ.Literal — Type

Literal <: Tree{<:Operator, <:Atom, 1}

A proposition represented by an atomic formula or its negation.

See also Operator, Atom, and Tree.

Examples

julia> @atomize PAndQ.Literal(𝒾, p)
p

julia> @atomize PAndQ.Literal(¬, p)
¬p

PAndQ.Clause — Type

Clause{AO <: AndOr, L <: Literal} <: Compound
Clause(::AO, ps = Literal[])
Clause(::AO, p::Proposition)
Clause(::Union{NullaryOperator, Atom, Literal})

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty Clause is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also Atom, Literal, AndOr, and NullaryOperator.

Examples

julia> PAndQ.Clause(∧)
⊤

julia> @atomize PAndQ.Clause(p)
p

julia> @atomize PAndQ.Clause(∨, [¬p, q])
¬p ∨ q

PAndQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Compound
Normal(::typeof(and), ps = Clause{typeof(or)}[])
Normal(::typeof(or), ps = Clause{typeof(and)}[])
Normal(::AO, ::Proposition)
Normal(::Union{NullaryOperator, Proposition})

A Proposition represented in conjunctive or disjunctive normal form.

Info

An empty Normal is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also Clause, AndOr, and NullaryOperator.

Examples

julia> PAndQ.Normal(⊤)
⊤

julia> @atomize PAndQ.Normal(∧, p ⊻ q)
(p ∨ q) ∧ (¬p ∨ ¬q)

AbstractTrees.jl

AbstractTrees.children — Function

children(::Proposition)

Return an iterator over the child nodes of the given Proposition.

Examples

julia> @atomize PAndQ.children(p)
()

julia> @atomize PAndQ.children(¬p)
(PAndQ.Variable(:p),)

julia> @atomize PAndQ.children(p ∧ q)
(PAndQ.Variable(:p), PAndQ.Variable(:q))

AbstractTrees.nodevalue — Function

nodevalue(::Union{Tree{O}, Clause{O}, Normal{O}}) where O

Return O.instance.

See also Compound.

Examples

julia> @atomize PAndQ.nodevalue(¬p)
not (generic function with 5 methods)

julia> @atomize PAndQ.nodevalue(p ∧ q)
and (generic function with 17 methods)

AbstractTrees.printnode — Function

printnode(::IO, ::Union{NullaryOperator, Proposition}; kwargs...)

See also Proposition.

Examples

julia> @atomize PAndQ.printnode(stdout, p)
p
julia> @atomize PAndQ.printnode(stdout, ¬p)
¬
julia> @atomize PAndQ.printnode(stdout, p ∧ q)
∧

AbstractTrees.NodeType — Type

NodeType(::Type{<:Atom})

See also Atom.

AbstractTrees.nodetype — Function

nodetype(::Type{<:Atom})

See also Atom.

Utility

PAndQ.child — Function

child(x)

Equivalent to only ∘ children

See also children

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as a Variable if it is undefined in the caller's scope. If isexpr(x, :$), return an expression that instantiates it as a Constant. If x is another expression, traverse it with recursive calls to atomize Otherise, return x.

Printing

PAndQ.symbol_of — Function

symbol_of(::Operator)

Return the Unicode symbol of the given Operator.

Examples

julia> PAndQ.symbol_of(⊤)
:⊤

julia> PAndQ.symbol_of(¬)
:¬

julia> PAndQ.symbol_of(∧)
:∧

PAndQ.parenthesize — Function

parenthesize(::IO, p)

PAndQ.print_node — Function

print_node(io, p)

PAndQ.show_atom — Function

show_atom(::IO, ::Atom)

See also Atom.

PAndQ.base_type — Function

base_type(::Proposition)

Semantics

PAndQ.interpret_bool — Function

interpret_bool

PAndQ.process_valuations — Function

process_valuations(valuations, p, f)

PAndQ.neutral_operator — Function

neutral_operator(::NullaryOperator)

Return a subtype of AndOr that is the neutral element of the given NullaryOperator.

Examples

julia> PAndQ.neutral_operator(⊤)
and (generic function with 17 methods)

julia> PAndQ.neutral_operator(⊥)
or (generic function with 17 methods)

PAndQ.eval_doubles — Function

eval_doubles(f, doubles)

Base.convert — Method

convert(::Type{<:Proposition}, ::Union{NullaryOperator, Proposition})

See also NullaryOperator and Proposition.