Internals

PAndQ.union_typeof — Function

union_typeof(xs)

PicoSAT

PAndQ.PicoSAT — Module

PicoSAT

This module provides an interface to libpicosat_jll.jl.

libpicosat

PAndQ.PicoSAT.picosat_init — Function

picosat_init()

Construct a new PicoSAT instance and return a pointer to it.

PAndQ.PicoSAT.picosat_reset — Function

picosat_reset(pico_sat)

Destruct the pico_sat instance.

PAndQ.PicoSAT.picosat_add — Function

picosat_add(pico_sat, literal)

Append the literal to the pico_sat instance's current disjunctive clause.

PAndQ.PicoSAT.picosat_variables — Function

picosat_variables(pico_sat)

Return the number of unique atoms in the pico_sat instance.

PAndQ.PicoSAT.picosat_print — Function

picosat_print(pico_sat, file)

Write the DIMACS format of the pico_sat instance's proposition to the given file.

PAndQ.PicoSAT.picosat_sat — Function

picosat_sat(pico_sat, limit)

Search for a satisfiable assignment of the pico_sat instance's proposition and return 0, 10, or 20 if the status is unknown, satisfiable, or unsatisfiable, respectively.

PAndQ.PicoSAT.picosat_deref — Function

picosat_deref(pico_sat, atom)

Return the assignment of the atom, where 1, -1, and 0 indicate true, false, and unknown, respectively.

This function must be called after verifying that the status of picosat_sat is satisfiable.

Utilities

PAndQ.PicoSAT.add_clause — Function

add_clause(pico_sat, clause)

Modify the pico_sat instance's proposition to be the conjunction of itself and the disjunctive clause.

PAndQ.PicoSAT.initialize — Function

initialize(clauses)

Return a PicoSAT pointer with its proposition being a conjunction of the disjunctive clauses.

PAndQ.PicoSAT.finalize! — Function

finalize!(solutions)

If the argument has not been finalized, call picosat_reset on its PicoSAT pointer and then set the pointer equal to C_NULL.

PAndQ.PicoSAT.Solutions — Type

Solutions
Solutions(clauses)

A stateful iterator of valuations that satisfy the given proposition.

Base.eltype — Function

eltype(::Type{Solutions})

Examples

julia> eltype(PAndQ.PicoSAT.Solutions)
Base.Generator{Base.Iterators.Enumerate{Base.Iterators.Filter{Base.Fix2{typeof(!=), Int64}, Vector{Int32}}}, Base.Splat{typeof(*)}}

Base.IteratorSize — Type

IteratorSize(::Type{Solutions})

Since counting the number of Solutions to a proposition is intractable, its IteratorSize is Base.SizeUnknown.

Examples

julia> Base.IteratorSize(PAndQ.PicoSAT.Solutions)
Base.SizeUnknown()

PAndQ.PicoSAT.is_satisfiable — Function

is_satisfiable(pico_sat)

Base.isdone — Function

isdone(solutions::Solutions, pico_sat = solutions.pico_sat)

Base.iterate — Function

iterate(solutions::Solutions, pico_sat = solutions.pico_sat)

If the status of pico_sat is_satisfiable, return a Tuple of the current solution and pico_sat. Otherwise, finalize! the solutions and return nothing.

PAndQ.PicoSAT.dimacs — Function

dimacs(io, clauses)

Examples

julia> PAndQ.PicoSAT.dimacs(stdout, ((-1, -2), (1, 2)))
p cnf 2 2
-1 -2 0
1 2 0

julia> PAndQ.PicoSAT.dimacs(String, ((1, -2), (-1, 2)))
"p cnf 2 2\n1 -2 0\n-1 2 0\n"

Interface

PAndQ.Interface.InterfaceError — Type

InterfaceError{F, T} <: Exception
InterfaceError(::F, ::T)

Base.showerror — Function

showerror(::IO, ::InterfaceError)

PAndQ.Interface.@interface — Macro

@interface(f, xs...)

Operators

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of Nullary Operators.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of Unary Operators.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of Binary Operators.

PAndQ.NaryOperator — Type

NaryOperator

The Union of Nary Operators.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

Types

Proposition
├─ Atom
│  ├─ Constant
│  └─ Variable
└─ Compound
   ├─ Clause
   ├─ Normal
   └─ Tree

Abstract

PAndQ.Proposition — Type

Proposition

A proposition.

Supertype of Atom and Compound.

PAndQ.Atom — Type

Atom <: Proposition

A proposition with no deeper propositional structure.

Subtype of Proposition. Supertype of Constant and Variable.

PAndQ.Compound — Type

Compound <: Proposition

A proposition composed from connecting Atoms with one or more Operators.

Subtype of Proposition. Supertype of Tree, Clause, and Normal.

Concrete

PAndQ.Constant — Type

Constant{T} <: Atom
Constant(::T)

An atomic sentence.

Subtype of Atom.

Examples

julia> PAndQ.Constant(1)
$(1)

julia> PAndQ.Constant("Logic is fun")
$("Logic is fun")

PAndQ.Variable — Type

Variable <: Atom

An atomic formula.

Subtype of Atom.

Examples

julia> PAndQ.Variable(:p)
p

julia> PAndQ.Variable(:q)
q

PAndQ.Tree — Type

Tree{O <: Operator, AT <: Union{Atom, Tree}, N} <: Compound
Tree(::NullaryOperator, ::Atom)
Tree(::Operator, ::Tree...)
Tree(::Proposition)

A Proposition represented by an abstract syntax tree.

Subtype of Compound. See also Atom, NullaryOperator, and Operator.

Examples

julia> PAndQ.Tree(⊤)
⊤

julia> @atomize PAndQ.Tree(¬, p)
¬p

julia> @atomize PAndQ.Tree(and, PAndQ.Tree(p), PAndQ.Tree(q))
p ∧ q

PAndQ.Literal — Type

Literal <: Tree{<:Operator, <:Atom, 1}

A proposition represented by an atomic formula or its negation.

See also Operator, Atom, and Tree.

Examples

julia> @atomize PAndQ.Literal(𝒾, p)
p

julia> @atomize PAndQ.Literal(¬, p)
¬p

PAndQ.Clause — Type

Clause{AO <: AndOr, A <: AbstractVector{<:Atom}, L <: AbstractVector{Int}} <: Compound
Clause(::AO, ::A, ::L)

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty Clause is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also Atom, Literal, AndOr, and NullaryOperator.

Examples

julia> PAndQ.Clause(∧, PAndQ.Atom[], Set{Int}())
⊤

julia> @atomize PAndQ.Clause(∧, [p], Set(1))
p

julia> @atomize PAndQ.Clause(∨, [p, q], Set((1, -2)))
¬q ∨ p

PAndQ.Normal — Type

Normal{AO <: AndOr, A <: Atom, C <: AbstractVector{<:AbstractVector{Int}}} <: Compound
Normal(::AO, ::A, ::C)

A Proposition represented in conjunctive or disjunctive normal form.

Info

An empty Normal is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also Clause, AndOr, and NullaryOperator.

Examples

julia> PAndQ.Normal(∧, PAndQ.Atom[], Set{Set{Int}}())
⊤

julia> @atomize PAndQ.Normal(∧, [p, q], Set(map(Set, ((1, 2), (-1, -2)))))
(¬p ∨ ¬q) ∧ (p ∨ q)

AbstractTrees.jl

AbstractTrees.children — Function

children(::Proposition)

Return an iterator over the child nodes of the given Proposition.

Examples

julia> @atomize PAndQ.children(p)
()

julia> @atomize PAndQ.children(¬p)
(PAndQ.Variable(:p),)

julia> @atomize PAndQ.children(p ∧ q)
(PAndQ.Variable(:p), PAndQ.Variable(:q))

AbstractTrees.nodevalue — Function

nodevalue(::Union{Tree{O}, Clause{O}, Normal{O}}) where O

Return O.instance.

See also Compound.

Examples

julia> @atomize PAndQ.nodevalue(¬p)
¬

julia> @atomize PAndQ.nodevalue(p ∧ q)
∧

AbstractTrees.printnode — Function

printnode(::IO, ::Union{NullaryOperator, Proposition}; kwargs...)

See also Proposition.

Examples

julia> @atomize PAndQ.printnode(stdout, p)
p
julia> @atomize PAndQ.printnode(stdout, ¬p)
¬
julia> @atomize PAndQ.printnode(stdout, p ∧ q)
∧

AbstractTrees.NodeType — Type

NodeType(::Type{<:Atom})

See also Atom.

AbstractTrees.nodetype — Function

nodetype(::Type{<:Atom})

See also Atom.

Utility

PAndQ.child — Function

child(x)

Equivalent to only ∘ children

See also children.

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as a Variable if it is undefined in the caller's scope. If isexpr(x, :$), return an expression that instantiates it as a Constant. If x is a different expression, traverse it with recursive calls to atomize. Otherise, return x.

PAndQ.distribute — Function

distribute(p)

PAndQ.flatten — Function

flatten(p)

Printing

Semantics

PAndQ.eval_doubles — Function

eval_doubles(f, doubles)

Base.convert — Method

convert(::Type{<:Proposition}, p)

See also Proposition.

Base.promote_rule — Function

promote_rule