Internals

Everything that is not exported is considered internal.

PAndQ.union_typeof — Function

union_typeof

Operators

Union Types

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of logical operators that take zero arguments.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of logical unary operators.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of logical binary operators.

PAndQ.LogicalOperator — Type

LogicalOperator

The Union of logical operators.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

Types

Proposition
├─ Atom
│  ├─ Constant
│  └─ Variable
└─ Compound
   ├─ Clause
   ├─ Expressive
   │  ├─ Normal
   │  └─ Tree
   └─ Literal

Abstract

PAndQ.Proposition — Type

Proposition

A logical proposition.

Supertype of Atom and Compound.

PAndQ.Atom — Type

Atom <: Proposition

A proposition with no deeper propositional structure.

Subtype of Proposition. Supertype of Constant and Variable.

PAndQ.Compound — Type

Compound{LO} <: Proposition

A proposition composed from connecting atomic propositions with logical operators.

Subtype of Proposition. Supertype of Literal, Clause, and Expressive.

PAndQ.Expressive — Type

Expressive{LO} <: Compound{LO}

A proposition that is expressively complete.

Subtype of Compound. Supertype of Tree and Normal.

Concrete

PAndQ.Constant — Type

Constant{T} <: Atom
Constant(value::T)

An atomic sentence.

Subtype of Atom.

Examples

julia> PAndQ.Constant(1)
$(1)

julia> PAndQ.Constant("Logic is fun")
$("Logic is fun")

PAndQ.Variable — Type

Variable <: Atom

An atomic formula.

Subtype of Atom.

Examples

julia> PAndQ.Variable(:p)
p

julia> PAndQ.Variable(:q)
q

PAndQ.Literal — Type

Literal{UO <: UnaryOperator, A <: Atom} <: Compound{UO}
Literal(::UO, ::A)
Literal(::Union{Atom, Literal})

A proposition represented by an atomic formula or its negation.

Subtype of Compound. See also UnaryOperator and Atom.

Examples

julia> @atomize PAndQ.Literal(𝒾, p)
p

julia> @atomize PAndQ.Literal(¬, p)
¬p

PAndQ.Tree — Type

Tree{LO <: LogicalOperator, AT <: Union{Atom, Tree}} <: Expressive{LO}
Tree(::NullaryOperator, ::Atom)
Tree(::LogicalOperator, ::Tree...)
Tree(::Proposition)

A proposition represented by an abstract syntax tree.

Subtype of Expressive. See also LogicalOperator.

Examples

julia> PAndQ.Tree(⊤)
⊤

julia> @atomize PAndQ.Tree(¬, p)
¬p

julia> @atomize PAndQ.Tree(and, PAndQ.Tree(p), PAndQ.Tree(q))
p ∧ q

PAndQ.Clause — Type

Clause{AO <: AndOr, L <: Literal} <: Compound{AO}
Clause(::AO, ps = Literal[])
Clause(::AO, p::Proposition)
Clause(::Union{NullaryOperator, Atom, Literal})

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty Clause is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also AndOr, Literal, and NullaryOperator.

Examples

julia> PAndQ.Clause(∧)
⊤

julia> @atomize PAndQ.Clause(p)
p

julia> @atomize PAndQ.Clause(∨, [¬p, q])
¬p ∨ q

PAndQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Expressive{AO}
Normal(::typeof(and), ps = Clause{typeof(or)}[])
Normal(::typeof(or), ps = Clause{typeof(and)}[])
Normal(::AO, ::Proposition)
Normal(::Union{NullaryOperator, Proposition})

A proposition represented in conjunctive or disjunctive normal form.

Info

An empty Normal is logically equivalent to the neutral element of it's binary operator.

Subtype of Expressive. See also AndOr, Clause, NullaryOperator, and Proposition.

Examples

julia> PAndQ.Normal(⊤)
⊤

julia> @atomize PAndQ.Normal(∧, p ⊻ q)
(p ∨ q) ∧ (¬p ∨ ¬q)

AbstractTrees.jl

AbstractTrees.children — Function

children(::Proposition)

Return an iterator over the child nodes of the given Proposition.

Examples

julia> @atomize PAndQ.children(p)
()

julia> @atomize PAndQ.children(¬p)
(PAndQ.Variable(:p),)

julia> @atomize PAndQ.children(p ∧ q)
2-element Vector{PAndQ.Tree{typeof(identity), PAndQ.Variable}}:
 p
 q

AbstractTrees.nodevalue — Function

nodevalue(::Compound)

See also Compound.

Examples

julia> @atomize PAndQ.nodevalue(¬p)
not (generic function with 19 methods)

julia> @atomize PAndQ.nodevalue(p ∧ q)
and (generic function with 18 methods)

AbstractTrees.printnode — Function

printnode(::IO, ::Proposition; kwargs...)

See also Proposition.

Examples

julia> @atomize PAndQ.printnode(stdout, p)
p
julia> @atomize PAndQ.printnode(stdout, ¬p)
¬
julia> @atomize PAndQ.printnode(stdout, p ∧ q)
∧

Utility

PAndQ.child — Function

child(x)

Equivalent to only(children(x))

See also children

PAndQ.union_all_type — Function

union_all_type(::Proposition)

Return the UnionAll type of a Proposition.

Examples

julia> @atomize PAndQ.union_all_type(p)
PAndQ.Variable

julia> @atomize PAndQ.union_all_type(p ∧ q)
PAndQ.Tree

PAndQ.symbol_value — Function

symbol_value

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as a Variable if it is undefined in the caller's scope. If isexpr(x, :$), return an expression that instantiates it as a Constant. If x is another expression, traverse it with recursive calls to atomize Otherise, return x.

Printing

PAndQ.alias_of — Function

alias_of(::LogicalOperator)

Return the written name of the given LogicalOperator.

PAndQ.symbol_of — Function

symbol_of(::LogicalOperator)

Return the Unicode symbol of the given LogicalOperator.

Examples

julia> PAndQ.symbol_of(⊤)
:⊤

julia> PAndQ.symbol_of(¬)
:¬

julia> PAndQ.symbol_of(∧)
:∧

PAndQ.parenthesize — Function

parenthesize(::IO, p)

PAndQ.print_node — Function

print_node(io, p)

PAndQ.show_atom — Function

show_atom(io, ::Atom)

Semantics

PAndQ.neutral_operator — Function

neutral_operator(::NullaryOperator)

Return a subtype of AndOr that is the neutral element of the given NullaryOperator.

See also left_neutrals and right_neutrals.

Examples

julia> PAndQ.neutral_operator(⊤)
and (generic function with 18 methods)

julia> PAndQ.neutral_operator(⊥)
or (generic function with 18 methods)

PAndQ.eval_doubles — Function

eval_doubles(f, doubles)