Internals

PAndQ.union_typeof — Function

union_typeof(xs)

Equivalent to Union{map(typeof, xs)...}.

Examples

julia> PAndQ.union_typeof((⊤, ¬, ∧))
Union{PAndQ.Interface.Operator{:tautology}, PAndQ.Interface.Operator{:not}, PAndQ.Interface.Operator{:and}}

Solvers

Z3

PAndQ.Z3 — Module

Z3

This module provides an interface to z3_jll.jl.

PAndQ.Z3.Library — Module

Library

PAndQ.Z3.add_clause — Function

add_clause(context, solver, clause)

PAndQ.Z3.Solutions — Type

Solutions

Base.IteratorSize — Method

IteratorSize(::Type{Solutions})

Examples

julia> Base.IteratorSize(PAndQ.Z3.Solutions)
Base.SizeUnknown()

Base.eltype — Method

eltype(::Type{Solutions})

The type of the elements generated by a Solutions iterator.

Examples

julia> eltype(PAndQ.Z3.Solutions)
Vector{Bool} (alias for Array{Bool, 1})

Base.isdone — Method

isdone

Base.iterate — Method

iterate

PicoSAT

PAndQ.PicoSAT — Module

PicoSAT

This module provides an interface to libpicosat_jll.jl.

Library

PAndQ.PicoSAT.picosat_init — Function

picosat_init()

Construct a new PicoSAT instance and return a pointer to it.

PAndQ.PicoSAT.picosat_reset — Function

picosat_reset(pico_sat)

Destruct the pico_sat instance.

PAndQ.PicoSAT.picosat_add — Function

picosat_add(pico_sat, literal)

Append the literal to the pico_sat instance's current disjunctive clause.

PAndQ.PicoSAT.picosat_adjust — Function

picosat_adjust

PAndQ.PicoSAT.picosat_variables — Function

picosat_variables(pico_sat)

Return the number of unique atoms in the pico_sat instance.

PAndQ.PicoSAT.picosat_print — Function

picosat_print(pico_sat, file)

Write the DIMACS format of the pico_sat instance's proposition to the given file.

PAndQ.PicoSAT.picosat_sat — Function

picosat_sat(pico_sat, limit)

Search for a satisfiable assignment of the pico_sat instance's proposition and return 0, 10, or 20 if the status is unknown, satisfiable, or unsatisfiable, respectively.

PAndQ.PicoSAT.picosat_deref — Function

picosat_deref(pico_sat, atom)

Return the assignment of the atom, where 1, -1, and 0 indicate true, false, and unknown, respectively.

This function must be called after verifying that the status of picosat_sat is satisfiable.

Utilities

PAndQ.PicoSAT.add_clause — Function

add_clause(pico_sat, clause)

Mutate the pico_sat instance's proposition to be the conjunction of itself and the disjunctive clause.

PAndQ.PicoSAT.initialize — Function

initialize(clauses, n)

Return a PicoSAT pointer with its proposition being a conjunction of the disjunctive clauses.

PAndQ.PicoSAT.Solutions — Type

Solutions
Solutions(clauses)

A stateful iterator of valuations that satisfy the given proposition.

Calling finalize on this iterator will first check whether it has already been finalized. If not, it will call picosat_reset on its PicoSAT pointer and set the pointer equal to C_NULL.

Base.eltype — Method

eltype(::Type{Solutions})

The type of the elements generated by a Solutions iterator.

Examples

julia> eltype(PAndQ.PicoSAT.Solutions)
Vector{Bool} (alias for Array{Bool, 1})

Base.IteratorSize — Method

IteratorSize(::Type{Solutions})

Since counting the number of Solutions to a proposition is intractable, its IteratorSize is Base.SizeUnknown.

Examples

julia> Base.IteratorSize(PAndQ.PicoSAT.Solutions)
Base.SizeUnknown()

Base.isdone — Method

isdone(solutions::Solutions, pico_sat = solutions.pico_sat)

Return a Boolean whether the pico_sat instance can yield any more solutions without advancing the Solutions iterator.

Finalize the iterator if it has not yet been finalized and is done.

Base.iterate — Method

iterate(solutions::Solutions, pico_sat = solutions.pico_sat)

If the status of pico_sat is satisfiable, return a Tuple of the current solution and pico_sat. Otherwise, finalize the solutions and return nothing.

PAndQ.PicoSAT.print_dimacs — Function

print_dimacs(io, clauses, n)

Print the DIMACS format of the conjunctive clauses.

Examples

julia> PAndQ.PicoSAT.print_dimacs(stdout, [[-1, -2], [1, 2]], 2)
p cnf 2 2
-1 -2 0
1 2 0

julia> PAndQ.PicoSAT.print_dimacs(stdout, [[1, -2], [-1, 2]], 2)
p cnf 2 2
1 -2 0
-1 2 0

Interface

PAndQ.Interface.InterfaceError — Type

InterfaceError{F <: Function, O <: Operator} <: Exception
InterfaceError(::F, ::O)

An Exception indicating that the function of type F has not been implemented for the value of type T.

Base.showerror — Function

showerror(::IO, ::InterfaceError)

Print a message indicating to implement a method of an interface.

PAndQ.Interface.@interface — Macro

@interface(f, xs...)

Define a fallback method that throws an InterfaceError.

Operators

PAndQ.NullaryOperator — Type

NullaryOperator

The Union of Nullary Operators.

PAndQ.UnaryOperator — Type

UnaryOperator

The Union of Unary Operators.

PAndQ.BinaryOperator — Type

BinaryOperator

The Union of Binary Operators.

PAndQ.NaryOperator — Type

NaryOperator

The Union of Nary Operators.

PAndQ.AndOr — Type

AndOr

The Union of and and or.

Propositions

Types

Proposition
├─ Atom
│  ├─ Constant
│  └─ Variable
└─ Compound
   └─ AbstractSyntaxTree

Abstract

PAndQ.Proposition — Type

Proposition

A proposition.

Supertype of Atom and Compound.

PAndQ.Atom — Type

Atom <: Proposition

A proposition with no deeper propositional structure.

Subtype of Proposition. Supertype of Constant and Variable.

PAndQ.Compound — Type

Compound <: Proposition

A proposition composed from connecting Atoms with one or more Operators.

Subtype of Proposition. Supertype of AbstractSyntaxTree.

Concrete

PAndQ.Constant — Type

Constant <: Atom
Constant(value)

An atomic sentence.

Subtype of Atom.

Examples

julia> PAndQ.Constant(1)
$(1)

julia> PAndQ.Constant("Logic is fun")
$("Logic is fun")

PAndQ.Variable — Type

Variable <: Atom

An atomic formula.

Subtype of Atom.

Examples

julia> PAndQ.Variable(:p)
p

julia> PAndQ.Variable(:q)
q

PAndQ.AbstractSyntaxTree — Type

AbstractSyntaxTree <: Compound
AbstractSyntaxTree(::Operator, ::Union{Atom, AbstractSyntaxTree})

A Proposition represented by an abstract syntax tree.

Subtype of Compound. See also Operator and Atom.

Examples

julia> PAndQ.AbstractSyntaxTree(⊤)
⊤

julia> @atomize PAndQ.AbstractSyntaxTree(¬, [p])
¬p

julia> @atomize PAndQ.AbstractSyntaxTree(and, [PAndQ.AbstractSyntaxTree(p), PAndQ.AbstractSyntaxTree(q)])
p ∧ q

AbstractTrees.jl

AbstractTrees.children — Function

children(::AbstractSyntaxTree)

Return an iterator over the child nodes of the given proposition.

See also AbstractSyntaxTree.

Examples

julia> @atomize PAndQ.children(p)
()

julia> @atomize PAndQ.children(¬p)
1-element Vector{PAndQ.Variable}:
 p

julia> @atomize PAndQ.children(p ∧ q)
2-element Vector{PAndQ.AbstractSyntaxTree}:
 p
 q

AbstractTrees.nodevalue — Function

nodevalue(::AbstractSyntaxTree)

Return the Operator of the proposition's root node.

See also AbstractSyntaxTree.

Examples

julia> @atomize PAndQ.nodevalue(¬p)
¬

julia> @atomize PAndQ.nodevalue(p ∧ q)
∧

AbstractTrees.printnode — Function

printnode(::IO, ::Union{Operator, Proposition}; kwargs...)

Print the representation of the proposition's root node.

See also Operator and Proposition.

Examples

julia> @atomize PAndQ.printnode(stdout, p)
p
julia> @atomize PAndQ.printnode(stdout, ¬p)
¬
julia> @atomize PAndQ.printnode(stdout, p ∧ q)
∧

AbstractTrees.NodeType — Type

NodeType(::Type{<:Atom})

See also Atom.

Examples

julia> PAndQ.NodeType(PAndQ.Atom)
AbstractTrees.HasNodeType()

AbstractTrees.nodetype — Function

nodetype(::Type{<:Atom})

See also Atom.

julia> PAndQ.nodetype(PAndQ.Atom)
PAndQ.Atom

Utility

PAndQ.deconstruct — Function

deconstruct(p)

Return (nodevalue(p), children(p)).

See also nodevalue and children.

Examples

julia> @atomize PAndQ.deconstruct(p)
(PAndQ.Variable(:p), ())

julia> @atomize PAndQ.deconstruct(¬p)
(not, PAndQ.Variable[PAndQ.Variable(:p)])

julia> @atomize PAndQ.deconstruct(p ∧ q)
(and, PAndQ.AbstractSyntaxTree[identical(PAndQ.Variable(:p)), identical(PAndQ.Variable(:q))])

PAndQ.child — Function

child(x)

Equivalent to only ∘ children.

See also children.

PAndQ.load_or_error — Function

load_or_error(x)

Return an expression that when evaluated returns x if PAndQ is defined and throws an Exception otherwise.

PAndQ.atomize — Function

atomize(x)

If x is a symbol, return an expression that instantiates it as a Variable if it is undefined in the caller's scope. If isexpr(x, :$), return an expression that instantiates it as a Constant. If x is a different expression, traverse it with recursive calls to atomize. Otherise, return x.

PAndQ.distribute — Function

distribute(p)

Given a proposition in negation normal form, return that proposition in conjunction normal form.

PAndQ.prune — Function

prune(p, atoms = Atom[], mapping = Dict{Atom, Int}())

PAndQ.reconstruct — Function

reconstruct(clauses, atoms)

Printing

Semantics

PAndQ.eval_pairs — Function

eval_pairs(f, pairs)

Define f(::typeof(left)) = right and f(::typeof(right)) = left for each pair left and right in pairs.

Base.convert — Function

convert(::Type{<:Proposition}, p)

See also Proposition.

Base.promote_rule — Function

promote_rule