Operators

Every possible truth table can be constructed with the functionally complete set of operators not and and. For convenience, all sixteen of them have been prepared. There are ten binary operators, with the remaining six being expressed with truth values, individual propositions, and the unary not operator.

julia> @p TruthTable(⊤, ⊥, ¬p, ¬q)
┌────────┬────────┬──────┬──────┬─────────┬─────────┐
│ ⊤      │ ⊥      │ p    │ q    │ ¬p      │ ¬q      │
│ Clause │ Clause │ Atom │ Atom │ Literal │ Literal │
├────────┼────────┼──────┼──────┼─────────┼─────────┤
│ ⊤      │ ⊥      │ ⊤    │ ⊤    │ ⊥       │ ⊥       │
│ ⊤      │ ⊥      │ ⊥    │ ⊤    │ ⊤       │ ⊥       │
├────────┼────────┼──────┼──────┼─────────┼─────────┤
│ ⊤      │ ⊥      │ ⊤    │ ⊥    │ ⊥       │ ⊤       │
│ ⊤      │ ⊥      │ ⊥    │ ⊥    │ ⊤       │ ⊤       │
└────────┴────────┴──────┴──────┴─────────┴─────────┘

Typing symbols with tab completion is performed by typing \, followed by the given characters, and then the [TAB] key. For example, ⊥ is typed with \bot[TAB]. See also Julia's documentation on Tab Completion and Unicode Input.

Operator associativity determines how operators with the same precedence group their operands. For example, ∧ is left associative. Therefore, p ∧ q ∧ r is equivalent to (p ∧ q) ∧ r. Operator precedence determines how expressions with distinct operators are grouped together. Higher precedence operators will group their operands before lower precedence operators. For example, ∧ has a higher precedence than ∨. Therefore, p ∨ q ∧ r is equivalent to p ∨ (q ∧ r), even though both operators are left associative. See also Julia's documentation on Operator Precedence and Associativity.

Info

== has a precedence of 7, which is higher than that of several binary operators. For some cases, you may need to use parentheses. For example, @p p → q == r will error, but @p (p → q) == r will correctly return false.

Name	Symbol	Tab completion	Associativity	Precedence

tautology	⊤	\top	none	0

contradiction	⊥	\bot	none	0

not	¬	\neg	right	0

and	∧	\wedge	left	12

nand	⊼	\nand	left	12

nor	⊽	\nor	left	11

or	∨	\vee	left	11

xor	⊻	\xor	left	11

xnor	↔	\leftrightarrow	right	4

imply	→	\rightarrow	right	4

not_imply	↛	\nrightarrow	right	4

converse_imply	←	\leftarrow	right	4

not_converse_imply	↚	\nleftarrow	right	4

<!– identity –>

PAQ.tautology — Function

⊤
⊤()
tautology
tautology()

Logical true operator.

⊤ can be typed by \top<tab>.

Examples

julia> ⊤()
tautology (generic function with 1 method)

julia> @p TruthTable(⊤)
┌────────┐
│ ⊤      │
│ Clause │
├────────┤
│ ⊤      │
└────────┘

PAQ.contradiction — Function

⊥
⊥()
contradiction
contradiction()

Logical false operator.

⊥ can be typed by \bot<tab>.

Examples

julia> ⊥()
contradiction (generic function with 1 method)

julia> @p TruthTable(⊥)
┌────────┐
│ ⊥      │
│ Clause │
├────────┤
│ ⊥      │
└────────┘

PAQ.not — Function

¬p
¬(p)
not(p)

Logical negation operator.

¬ can be typed by \neg<tab>.

Examples

julia> @p TruthTable(¬p)
┌──────┬─────────┐
│ p    │ ¬p      │
│ Atom │ Literal │
├──────┼─────────┤
│ ⊤    │ ⊥       │
│ ⊥    │ ⊤       │
└──────┴─────────┘

PAQ.and — Function

p ∧ q
∧(p, q)
and(p, q)

Logical conjunction operator.

∧ can be typed by \wedge<tab>.

Examples

julia> @p TruthTable(p ∧ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ∧ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

Base.nand — Function

p ⊼ q
⊼(p, q)
nand(p, q)

Logical non-conjunction operator.

⊼ can be typed by \nand<tab>.

Examples

julia> @p TruthTable(p ⊼ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ⊼ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

Base.nor — Function

p ⊽ q
⊽(p, q)
nor(p, q)

Logical non-disjunction operator.

⊽ can be typed by \nor<tab>.

Examples

julia> @p TruthTable(p ⊽ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ⊽ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAQ.or — Function

p ∨ q
∨(p, q)
or(p, q)

Logical disjunction operator.

∨ can be typed by \vee<tab>.

Examples

julia> @p TruthTable(p ∨ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ∨ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

Base.xor — Function

p ⊻ q
⊻(p, q)
xor(p, q)

Logical exclusive disjunction operator.

⊻ can be typed by \xor<tab>.

Examples

julia> @p TruthTable(p ⊻ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ⊻ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

PAQ.xnor — Function

p ↔ q
↔(p, q)
xnor(p, q)

Logical exclusive non-disjunction and biconditional operator.

↔ can be typed by \leftrightarrow<tab>.

Examples

julia> @p TruthTable(p ↔ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ↔ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAQ.imply — Function

p → q
→(p, q)
imply(p, q)

Logical implication operator.

→ can be typed by \rightarrow<tab>.

Examples

julia> @p TruthTable(p → q)
┌──────┬──────┬───────┐
│ p    │ q    │ p → q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAQ.not_imply — Function

p ↛ q
↛(p, q)
not_imply(p, q)

Logical non-implication operator.

↛ can be typed by \nrightarrow<tab>.

Examples

julia> @p TruthTable(p ↛ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ↛ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

PAQ.converse_imply — Function

p ← q
←(p, q)
converse_imply(p, q)

Logical converse implication operator.

← can be typed by \leftarrow<tab>.

Examples

julia> @p TruthTable(p ← q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ← q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊤     │
│ ⊥    │ ⊤    │ ⊥     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊤     │
│ ⊥    │ ⊥    │ ⊤     │
└──────┴──────┴───────┘

PAQ.not_converse_imply — Function

p ↚ q
↚(p, q)
not_converse_imply(p, q)

Logical converse non-implication operator.

↚ can be typed by \nleftarrow<tab>.

Examples

julia> @p TruthTable(p ↚ q)
┌──────┬──────┬───────┐
│ p    │ q    │ p ↚ q │
│ Atom │ Atom │ Tree  │
├──────┼──────┼───────┤
│ ⊤    │ ⊤    │ ⊥     │
│ ⊥    │ ⊤    │ ⊤     │
├──────┼──────┼───────┤
│ ⊤    │ ⊥    │ ⊥     │
│ ⊥    │ ⊥    │ ⊥     │
└──────┴──────┴───────┘

PAQ.reduce_and — Function

⋀(ps)
reduce_and(ps)

Equivalent to reduce(and, ps).

⋀ can be typed by \bigwedge<tab>.

Examples

PAQ.reduce_or — Function

⋁(ps)
reduce_or(ps)

Equivalent to reduce(and, ps).

⋁ can be typed by \bigvee<tab>.

Examples