Pretty Printing

AbstractTrees.print_tree — Function

print_tree(p, max_depth = typemax(Int64))

Prints a tree diagram of p.

If p isn't a Tree, it will be converted to one. The optional argument max_depth will truncate sub-trees at that depth.

julia> @p print_tree(p ⊻ q)
⊻
├─ p
└─ q

julia> @p print_tree((p ∧ ¬q) ∨ (¬p ∧ q))
∨
├─ ∧
│  ├─ p
│  └─ ¬
│     └─ q
└─ ∧
   ├─ ¬
   │  └─ p
   └─ q

PAQ.truth_table — Function

truth_table(::AbstractArray; numbered_rows = false)
truth_table(ps...; numbered_rows = false)

Print a truth table for the given Propositions and BinaryOperators.

If numbered_rows = true, the first column will contain each row's sequential number.

The first row of the header is the expression representing that column's proposition, while the second row indicates that expression's type. Logically equivalent propositions will be grouped in the same column, seperated by a comma.

See also tautology and contradiction.

Examples

julia> @p truth_table(p ∧ ¬p, p ∧ q)
┌────────┬──────┬──────┬───────┐
│ p ∧ ¬p │ p    │ q    │ p ∧ q │
│ Tree   │ Atom │ Atom │ Tree  │
├────────┼──────┼──────┼───────┤
│ ⊥      │ ⊤    │ ⊤    │ ⊤     │
│ ⊥      │ ⊥    │ ⊤    │ ⊥     │
├────────┼──────┼──────┼───────┤
│ ⊥      │ ⊤    │ ⊥    │ ⊥     │
│ ⊥      │ ⊥    │ ⊥    │ ⊥     │
└────────┴──────┴──────┴───────┘

julia> truth_table([⊻, imply], numbered_rows = true)
┌───┬──────┬──────┬────────┬────────┐
│ # │ _    │ __   │ _ ⊻ __ │ _ → __ │
│   │ Atom │ Atom │ Tree   │ Tree   │
├───┼──────┼──────┼────────┼────────┤
│ 1 │ ⊤    │ ⊤    │ ⊥      │ ⊤      │
│ 2 │ ⊥    │ ⊤    │ ⊤      │ ⊤      │
├───┼──────┼──────┼────────┼────────┤
│ 3 │ ⊤    │ ⊥    │ ⊤      │ ⊥      │
│ 4 │ ⊥    │ ⊥    │ ⊥      │ ⊤      │
└───┴──────┴──────┴────────┴────────┘

PAQ.@truth_table — Macro

@truth_table(ps...; numbered_rows = false)

Equivalent to @p truth_table(ps...; numbered_rows = false).

See also @p and truth_table.

Examples

julia> @truth_table ¬p Clause(and, p, q)
┌──────┬──────┬─────────┬────────┐
│ p    │ q    │ ¬p      │ p ∧ q  │
│ Atom │ Atom │ Literal │ Clause │
├──────┼──────┼─────────┼────────┤
│ ⊤    │ ⊤    │ ⊥       │ ⊤      │
│ ⊥    │ ⊤    │ ⊤       │ ⊥      │
├──────┼──────┼─────────┼────────┤
│ ⊤    │ ⊥    │ ⊥       │ ⊥      │
│ ⊥    │ ⊥    │ ⊤       │ ⊥      │
└──────┴──────┴─────────┴────────┘

julia> @truth_table (⊻) imply numbered_rows = true
┌───┬──────┬──────┬────────┬────────┐
│ # │ _    │ __   │ _ ⊻ __ │ _ → __ │
│   │ Atom │ Atom │ Tree   │ Tree   │
├───┼──────┼──────┼────────┼────────┤
│ 1 │ ⊤    │ ⊤    │ ⊥      │ ⊤      │
│ 2 │ ⊥    │ ⊤    │ ⊤      │ ⊤      │
├───┼──────┼──────┼────────┼────────┤
│ 3 │ ⊤    │ ⊥    │ ⊤      │ ⊥      │
│ 4 │ ⊥    │ ⊥    │ ⊥      │ ⊤      │
└───┴──────┴──────┴────────┴────────┘