Printing

PAndQ.TruthTable — Type

TruthTable(ps)

Construct a truth table for the given Propositions and LogicalOperators.

The header is a vector containing vectors of logically equivalent propositions. The sub_header corresponds to the header, but contains each proposition's UnionAll type. The body is a matrix where the rows contain interpretations and the columns correspond to elements in the header and sub_header.

Logically equivalent propositions are grouped in the same column.

See also tautology and contradiction.

Examples

julia> @p TruthTable(p ∧ ¬p, p ∧ q)
┌────────┬──────┬──────┬───────┐
│ p ∧ ¬p │ p    │ q    │ p ∧ q │
│ Tree   │ Atom │ Atom │ Tree  │
├────────┼──────┼──────┼───────┤
│ ⊥      │ ⊤    │ ⊤    │ ⊤     │
│ ⊥      │ ⊥    │ ⊤    │ ⊥     │
├────────┼──────┼──────┼───────┤
│ ⊥      │ ⊤    │ ⊥    │ ⊥     │
│ ⊥      │ ⊥    │ ⊥    │ ⊥     │
└────────┴──────┴──────┴───────┘

julia> TruthTable([⊻, imply])
┌──────┬──────┬────────┬────────┐
│ _    │ __   │ _ ⊻ __ │ _ → __ │
│ Atom │ Atom │ Tree   │ Tree   │
├──────┼──────┼────────┼────────┤
│ ⊤    │ ⊤    │ ⊥      │ ⊤      │
│ ⊥    │ ⊤    │ ⊤      │ ⊤      │
├──────┼──────┼────────┼────────┤
│ ⊤    │ ⊥    │ ⊤      │ ⊥      │
│ ⊥    │ ⊥    │ ⊥      │ ⊤      │
└──────┴──────┴────────┴────────┘

PAndQ.print_truth_table — Function

print_truth_table([io::Union{IO, String}], x, backend = :text, kwargs...)

Examples

PAndQ.println_truth_table — Function

println_truth_table(args...; kwargs...)

Equivalent to print_truth_table(args...; kwargs..., newline = true).

Examples

PAndQ.print_tree — Function

print_tree([io::Union{IO, String} = stdout], p; max_depth = typemax(Int64), newline = false, kwargs...)

Prints a tree diagram of p.

If p isn't a Tree, it will be converted to one. The optional argument max_depth will truncate sub-trees at that depth.

julia> @p print_tree(p ⊻ q)
⊻
├─ p
└─ q

julia> @p print_tree((p ∧ ¬q) ∨ (¬p ∧ q))
∨
├─ ∧
│  ├─ p
│  └─ ¬
│     └─ q
└─ ∧
   ├─ ¬
   │  └─ p
   └─ q

PAndQ.println_tree — Function

println_tree(args...; kwargs...)

Equivalent to print_tree(args...; kwargs..., newline = true).

Examples

PAndQ.print_latex — Function

print_latex([io::Union{IO, String} = stdout], x, delimeter = "\(" => "\)")

Return a string representation of x enclosed by delimeter, replacing each symbol with it's respective command.

Examples

julia> @p s = print_latex(String, p ∧ q)
"\\(p \\wedge q\\)"

julia> println(s)
\(p \wedge q\)

PAndQ.println_latex — Function

println_latex(args...; kwargs...)

Equivalent to print_latex(args...; kwargs..., newline = true).

Examples

PAndQ.print_markdown — Function

print_markdown

Examples

PAndQ.println_markdown — Function

println_markdown(args...; kwargs...)

Equivalent to print_markdown(args...; kwargs..., newline = true).

Examples