Propositional Logic

PAQ.Primitive — Type

Primitive <: Language
Primitive([statement::String = ""])

Primitive proposition.

Subtype of Language. See also Compound.

Examples

julia> p
Primitive("p")

julia> p()
2-element Vector{Pair{Vector{Pair{Primitive, Truth}}}}:
 [Primitive("p") => ⊤] => ⊤
 [Primitive("p") => ⊥] => ⊥

PAQ.Boolean — Type

Boolean <: Operator

Set of functionally complete logical connectives.

Subtype of Operator. Supertype of Not and And. See also Boolean Operators.

PAQ.Not — Type

Not <: Boolean <: Operator

Singleton type representing logical negation.

Subtype of Boolean and Operator. See also And and Language.

Examples

julia> @truth_table PAQ.Not()(p)
┌───────────┬────────────────┐
│ p         │ (PAQ.Not())(p) │
│ Primitive │ Propositional  │
│ "p"       │                │
├───────────┼────────────────┤
│ ⊤         │ ⊥              │
│ ⊥         │ ⊤              │
└───────────┴────────────────┘

PAQ.And — Type

And <: Boolean <: Operator

Singleton type representing logical conjunction.

Subtype of Boolean and Operator. See also Not and Language.

Examples

julia> @truth_table PAQ.And()(p, q)
┌───────────┬───────────┬───────────────────┐
│ p         │ q         │ (PAQ.And())(p, q) │
│ Primitive │ Primitive │ Propositional     │
│ "p"       │ "q"       │                   │
├───────────┼───────────┼───────────────────┤
│ ⊤         │ ⊤         │ ⊤                 │
│ ⊤         │ ⊥         │ ⊥                 │
├───────────┼───────────┼───────────────────┤
│ ⊥         │ ⊤         │ ⊥                 │
│ ⊥         │ ⊥         │ ⊥                 │
└───────────┴───────────┴───────────────────┘

PAQ.Propositional — Type

Propositional{
    L <: Union{
        Primitive,
        Tuple{Not, Compound},
        Tuple{And, Compound, Compound}
    }
}(ϕ::L) <: Compound <: Language

Abstract syntax tree representing a compound proposition.

Subtype of Compound and Language. See also Primitive, Not, and And.

Examples

julia> p ∧ ¬p
Propositional(
  And(), Propositional(
    Primitive("p")
  ) Propositional(
    Not(), Propositional(
      Primitive("p")
    ) 
  ) 
)

julia> (p ∧ ¬p)()
2-element Vector{Pair{Vector{Pair{Primitive, Truth}}, Truth{Val{:⊥}}}}:
 [Primitive("p") => ⊤] => ⊥
 [Primitive("p") => ⊥] => ⊥

julia> (p → q) ∧ (p ← q) == ¬(p ⊻ q)
true

PAQ.Truth — Type

Truth{V <: Union{Val{:⊥}, Val{:⊤}}}(::V) <: Language

Container for the constants tautology and contradiction. Subtype of Language.

PAQ.tautology — Constant

⊤
tautology

A constant which is true in every possible interpretation.

One of two valid instances of Truth, the other instance being contradiction.

⊤ can be typed by \top<tab>.

Examples

julia> ¬⊤
⊥

julia> tautology()
1-element Vector{Pair{Vector{Pair{Primitive, Truth}}, Truth{Val{:⊤}}}}:
 [] => ⊤

PAQ.contradiction — Constant

⊥
contradiction

A constant which is false in every possible interpretation.

One of two valid instances of Truth, the other instance being tautology.

⊥ can be typed by \bot<tab>.

Examples

julia> ¬⊥
⊤

julia> contradiction()
1-element Vector{Pair{Vector{Pair{Primitive, Truth}}, Truth{Val{:⊥}}}}:
 [] => ⊥

PAQ.@primitive — Macro

@primitive(ps...)

Instantiates Primitive propositions.

Examples

julia> @primitive p q

julia> p
Primitive("p")

julia> q
Primitive("q")

PAQ.get_primitives — Function

get_primitives(ps::Language...)

Returns a vector of Primitive propositions contained in p.

Note that some primitives may optimized out of an expression, such as in p ∧ ⊥.

See also Language.

Examples

julia> get_primitives(p)
1-element Vector{Primitive}:
 Primitive("p")

julia> get_primitives(p ∧ q, r)
3-element Vector{Primitive}:
 Primitive("p")
 Primitive("q")
 Primitive("r")