Propositional Logic

PAQ.Boolean — Type

Boolean <: Operator

Set of logical connectives.

Subtype of Operator. Supertype of Not, And, and Or. See also Boolean Operators.

PAQ.Not — Type

Not <: Boolean <: Operator

Singleton type representing logical negation.

Subtype of Boolean and Operator. See also not. ```

PAQ.And — Type

And <: Boolean <: Operator

Singleton type representing logical conjunction.

Subtype of Boolean and Operator. See also and. ```

PAQ.Or — Type

Or <: Boolean <: Operator

Singleton type representing logical disjunction.

Subtype of Boolean and Operator. See also or.

PAQ.Primitive — Type

Primitive <: Language
Primitive([statement::String = ""])

Primitive proposition.

Subtype of Language.

Examples

julia> p
Primitive:
  p

julia> p()
Contingency:
  [p => ⊤] => ⊤
  [p => ⊥] => ⊥

PAQ.Literal — Type

Literal{
    L <: Union{
        Primitive,
        Tuple{Not, Primitive}
    }
} <: Compound <: Language
Literal(p::L)

A Primitive or its negation.

Subtype of Compound and Language. See also Not.

Examples

julia> r = ¬p
Literal:
  ¬p

julia> r()
Contingency:
  [p => ⊤] => ⊥
  [p => ⊥] => ⊤

PAQ.Propositional — Type

Propositional{
    L <: Union{
        Tuple{Not, Compound},
        Tuple{And, Compound, Compound}
    }
} <: Compound <: Language
Propositional(ϕ::L)

Abstract syntax tree representing a compound proposition.

Subtype of Compound and Language. See also Not and And.

Examples

julia> r = p ∧ ¬p
Propositional:
  p ∧ ¬p

julia> r()
Truth:
  ⊥

julia> (p ∧ q)()
Contingency:
  [p => ⊤, q => ⊤] => ⊤
  [p => ⊤, q => ⊥] => ⊥
  [p => ⊥, q => ⊤] => ⊥
  [p => ⊥, q => ⊥] => ⊥

PAQ.Normal — Type

Normal{B <: Union{And, Or}} <: Compound <: Language
Normal(::B, clauses::Vector{Vector{Literal}})

The conjunctive or disjunctive normal form of a proposition.

Constructing an instance with the parameters (And(), p) and (Or(), p) correspond to conjunctive and disjunctive normal form, respectively.

Subtype of Compound and Language.

Examples

julia> r = Normal(And(), p ∧ q)
Normal:
  (¬p ∨ q) ∧ (p ∨ ¬q) ∧ (p ∨ q)

julia> s = Normal(Or(), ¬p ∨ ¬q)
Normal:
  (p ∧ ¬q) ∨ (¬p ∧ q) ∨ (¬p ∧ ¬q)

julia> t = r ∧ s
Propositional:
  (¬p ∨ q) ∧ (p ∨ ¬q) ∧ (p ∨ q) ∧ (p ∧ ¬q) ∨ (¬p ∧ q) ∨ (¬p ∧ ¬q)

julia> t()
Truth:
  ⊥

PAQ.Truth — Type

Truth{V <: Union{Val{:⊥}, Val{:⊤}}} <: Language
Truth(::V)

Container for the constants tautology and contradiction. Subtype of Language.

PAQ.tautology — Constant

⊤
tautology

A constant which is true in every possible interpretation.

One of two valid instances of Truth, the other instance being contradiction.

⊤ can be typed by \top<tab>.

Examples

julia> ¬⊤
Truth:
  ⊥

julia> tautology()
Truth:
  ⊤

PAQ.contradiction — Constant

⊥
contradiction

A constant which is false in every possible interpretation.

One of two valid instances of Truth, the other instance being tautology.

⊥ can be typed by \bot<tab>.

Examples

julia> ¬⊥
Truth:
  ⊤

julia> contradiction()
Truth:
  ⊥

PAQ.Contingency — Type

Contingency

Examples

julia> p()
Contingency:
  [p => ⊤] => ⊤
  [p => ⊥] => ⊥

julia> (p ∧ q)()
Contingency:
  [p => ⊤, q => ⊤] => ⊤
  [p => ⊤, q => ⊥] => ⊥
  [p => ⊥, q => ⊤] => ⊥
  [p => ⊥, q => ⊥] => ⊥

PAQ.@primitive — Macro

@primitive(ps...)

Instantiates Primitive propositions.

Examples

julia> @primitive p q

julia> p
Primitive:
  p

julia> q
Primitive:
  q

PAQ.get_primitives — Function

get_primitives(ps::Language...)

Returns a vector of Primitive propositions contained in ps.

Note that some primitives may optimized out of an expression, such as in p ∧ ⊥.

See also Language.

Examples

julia> r = p ∧ q
Propositional:
  p ∧ q

julia> get_primitives(r)
2-element Vector{Primitive}:
 Primitive("p")
 Primitive("q")