Propositions

Tip

Propositions can be converted into different, but logically equivalent forms (see also ==). For example, ⊥ == Tree(⊥) == Clause(⊥) == Normal(⊥). However, not all forms are Expressively complete. Otherwise, the conversion may throw an exception. For example, there is no way to represent Literal(⊥).

Proposition
├─ Atom
└─ Compound
   ├─ Clause
   ├─ Expressive
   │  ├─ Normal
   │  └─ Tree
   └─ Literal

PAndQ.Proposition — Type

Proposition

The set of well-formed logical formulae.

Supertype of Atom and Compound.

PAndQ.Compound — Type

Compound <: Proposition

A proposition composed from connecting Atomicpropositions with LogicalOperators.

Subtype of Proposition. Supertype of Literal, Clause, and Expressive.

PAndQ.Expressive — Type

Expressive <: Compound

A proposition that is expressively complete.

Subtype of Compound. Supertype of Tree and Normal.

PAndQ.Atom — Type

Atom{T} <: Proposition
Atom(::T)
Atom(::AtomicProposition)

A proposition with no deeper propositional structure.

Tip

Define pretty-printing for an instance of Atom{T} by overloading show(io::IO, p::Atom{T}).

Tip

Use @atoms and @p as shortcuts to define atoms or instantiate them inline, respectively.

Subtype of Proposition. See also AtomicProposition and LiteralProposition.

Examples

julia> Atom(:p)
p

julia> Atom("Logic is fun")
Atom("Logic is fun")

PAndQ.Literal — Type

Literal{UO <: UnaryOperator, T} <: Compound
Literal(::UO, ::Atom{T})
Literal(::LiteralProposition)

A proposition represented by an atomic formula or its negation.

Subtype of Compound. See also UnaryOperator, Atom, and LiteralProposition.

Examples

julia> r = @p ¬p
¬p

julia> ¬r
p

PAndQ.Tree — Type

Tree{
    LO <: LogicalOperator,
    P <: Union{Tuple{Proposition}, Tuple{Proposition, Proposition}}
} <: Expressive
Tree(::Union{NullaryOperator, Proposition})
Tree(::UnaryOperator, ::Atom)
Tree(::BinaryOperator, ::Tree, ::Tree)

A proposition represented by an abstract syntax tree.

Subtype of Expressive.

Examples

julia> @p r = p ⊻ q
p ⊻ q

julia> @p ¬r → s
(p ↔ q) → s

PAndQ.Clause — Type

Clause{AO <: AndOr, L <: Literal} <: Compound
Clause(::AO, ps = Literal[])
Clause(::AO, p::Proposition)
Clause(::Union{NullaryOperator, Proposition})

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty clause is logically equivalent to the neutral element of it's binary operator.

See also Literal and NullaryOperator. Subtype of Compound.

Examples

julia> Clause(and)
⊤

julia> @p Clause(p)
p

julia> @p Clause(or, [¬p, q])
¬p ∨ q

PAndQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Expressive
Normal(::typeof(and), ::Vector{Clause{typeof(or)}} = Clause{typeof(or)}[])
Normal(::typeof(or), ::Vector{Clause{typeof(and)}} = Clause{typeof(and)}[])
Normal(::AO, ::Proposition)
Normal(::Union{NullaryOperator, Proposition})

A proposition represented in conjunctive or disjunctive normal form.

Info

An empty normal form is logically equivalent to the neutral element of it's binary operator.

See also Clause and NullaryOperator. Subtype of Expressive.

Examples

julia> s = @p Normal(and, [Clause(or, [p, q]), Clause(or, ¬r)])
(p ∨ q) ∧ (¬r)

julia> ¬s
(¬p ∧ ¬q) ∨ (r)