Propositional Logic

Tip

Propositions can be converted into different, but logically equivalent forms (see also ==). For example, ⊥ == Valuation(⊥) == Tree(⊥) == Clause(⊥). However, not all forms are Expressively complete. Otherwise, the conversion may throw an exception. For example, there is no way to represent Literal(⊥).

Proposition
├─ Atom
└─ Compound
   ├─ Clause
   ├─ Expressive
   │  ├─ Normal
   │  ├─ Tree
   │  └─ Valuation
   └─ Literal

PAQ.Proposition — Type

Proposition

The set of well-formed logical formulae.

Supertype of Atom and Compound. ```

PAQ.Compound — Type

Compound <: Proposition

A proposition composed from one or more Atomic propositions.

Subtype of Proposition. Supertype of Literal, Clause, and Expressive.

PAQ.Expressive — Type

Expressive <: Compound

A proposition that is expressively complete.

Subtype of Compound. Supertype of Valuation, Tree, Normal, and Pretty.

PAQ.Atom — Type

Atom{SS <: Union{String, Symbol}} <: Proposition
Atom([::Union{Symbol, String])
Atom(::Proposition)

A proposition with no deeper propositional structure.

A string argument can be thought of as a specific statement, while a symbol can be variable. However, the only builtin difference between these are how they pretty-print. An atom with a string argument will be encompassed by quotation marks, while an atom with a symbol argument will only show the symbol's characters.

Tip

Using single character symbols (with corresponding variable names, if applicable) reproduces the syntax commonly found in logical literature.

Info

Atom() defaults to Atom(:_). This underscore symbol is useful as a default, such as when representing a truth value. For example, Tree(⊥) is pretty-printed as _ ∧ ¬_. This is a special case; it is not idiomatic to use for most purposes.

Subtype of Proposition.

Examples

julia> Atom(:p)
Atom:
 p

julia> Atom("Logic is fun")
Atom:
 "Logic is fun"

PAQ.Literal — Type

Literal{UO <: UnaryOperator} <: Compound
Literal(::UO, ::Atom)
Literal(::Proposition)

A proposition represented by an atomic formula or its negation.

Subtype of Compound. See also UnaryOperator and Atom.

Examples

julia> r = @p ¬p
Literal:
 ¬p

julia> ¬r
Atom:
 p

PAQ.Clause — Type

Clause{
    AO <: AndOr,
    L <: Literal
} <: Compound
Clause(::AO, [::Vector{Proposition}])
Clause(::AO, ::Proposition...)

A proposition represented as either a [conjunction or disjunction of literals](https://en.wikipedia.org/wiki/Clause_(logic).

Info

An empty clause is logically equivalent to the identity element of it's binary operator.

See also Literal. Subtype of Compound.

Examples

julia> Clause(and)
Clause:
 ⊥

julia> @p Clause(and, p, q)
Clause:
 p ∧ q

julia> @p Clause(or, [¬p, q])
Clause:
 ¬p ∨ q

PAQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Expressive
Normal(::typef(and), ::Vector{Clause{typeof(or)}})
Normal(::typef(or), ::Vector{Clause{typeof(and)}})
Normal(::Proposition)

A proposition represented in conjunctive or disjunctive normal form.

Info

An empty normal form is logically equivalent to the identity element of it's binary operator.

Subtype of Expressive.

Examples

julia> s = @p Normal(and, Clause(or, p, q), Clause(or, ¬r))
Normal:
 (p ∨ q) ∧ (¬r)

julia> ¬s
Normal:
 (¬p ∧ ¬q) ∨ (r)

PAQ.Valuation — Type

Valuation{P <: Pair} <: Expressive
Valuation(::Vector{P})
Valuation(::Proposition)

Proposition represented by a vector of interpretations.

Subtype of Expressive.

Examples

julia> @p Valuation(p ∧ q)
Valuation:
 [p => ⊤, q => ⊤] => ⊤
 [p => ⊥, q => ⊤] => ⊥
 [p => ⊤, q => ⊥] => ⊥
 [p => ⊥, q => ⊥] => ⊥

julia> Valuation(⊥)
Valuation:
 [] => ⊥

PAQ.Tree — Type

Tree{
    O <: BooleanOperator,
    P <: Union{Proposition, Tuple{Proposition, Proposition}}
} <: Expressive
Tree(::UnaryOperator, ::Atom)
Tree(::BinaryOperator, ::Tree, ::Tree)
Tree(::Proposition)

Proposition represented by an abstract syntax tree.

Subtype of Expressive.

Examples

julia> r = @p p ⊻ q
Tree:
 p ⊻ q

julia> @p ¬r → s
Tree:
 (p ↔ q) → s