Propositions

Macros

PAndQ.@atomize — Macro

@atomize(expression)

Instantiate Constants and Variables inline.

Constants are instantiated with the $ interpolation syntax. Variables are instantiated with previously undefined symbols.

Warning

This macro attempts to ignore symbols that are being assigned a value. For example, @atomize f(; x = p) = x ∧ q should be equivalent to f(; x = @atomize p)) = x ∧ @atomize q. However, this feature is in-progress and only works in some cases. The implementation is cautious to skip the parts of the expression that it cannot yet handle.

Examples

julia> @atomize x = p ∧ q
p ∧ q

julia> @atomize x → r
(p ∧ q) → r

julia> @atomize $1 ∧ $(1 + 1)
$(1) ∧ $(2)

PAndQ.@variables — Macro

@variables(ps...)

Define Variables and return a vector containing them.

Each symbol p is defined as p = PAndQ.Variable(:p).

Examples

julia> @variables p q
2-element Vector{PAndQ.Variable}:
 p
 q

julia> p
p

julia> q
q

Utilities

PAndQ.atoms — Function

atoms(p)

Return an iterator of each Atom contained in p.

Examples

julia> @atomize collect(atoms(p ∧ q))
2-element Vector{PAndQ.Variable}:
 p
 q

PAndQ.operators — Function

operators(p)

Return an iterator of each operator contained in p.

Examples

julia> @atomize collect(operators(¬p))
1-element Vector{typeof(not)}:
 not (generic function with 5 methods)

julia> @atomize collect(operators(¬p ∧ q))
3-element Vector{Function}:
 and (generic function with 17 methods)
 not (generic function with 5 methods)
 identity (generic function with 1 method)

PAndQ.value — Function

value(::Proposition)

Unwrap the value of a Constant.

The Proposition must be logically equivalent to a Constant.

Examples

julia> @atomize value($1)
1