Propositions

Instantiation

PAndQ.@atomize — Macro

@atomize(expression)

Instantiate Constants and Variables inline.

Constants are instantiated with the $ interpolation syntax. Variables are instantiated with previously undefined symbols.

Warning

This macro attempts to ignore symbols that are being assigned a value. For example, @atomize f(; x = p) = x ∧ q should be equivalent to f(; x = @atomize p)) = x ∧ @atomize q. However, this feature is in-progress and only works in some cases. The implementation is cautious to skip the parts of the expression that it cannot yet handle.

Examples

julia> @atomize x = p ∧ q
p ∧ q

julia> @atomize x → r
(p ∧ q) → r

julia> @atomize $1 ∧ $(1 + 1)
$(1) ∧ $(2)

PAndQ.@variables — Macro

@variables(ps...)

Define Variables and return a vector containing them.

Each symbol p is defined as p = PAndQ.Variable(:p).

Examples

julia> @variables p q
2-element Vector{PAndQ.Variable}:
 p
 q

julia> p
p

julia> q
q

PAndQ.constants — Function

constants(xs)

Instantiate each element as a Constant.

Examples

julia> constants(1:2)
2-element Vector{PAndQ.Constant{Int64}}:
 $(1)
 $(2)

julia> constants([[1,2], [3,4]])
2-element Vector{PAndQ.Constant{Vector{Int64}}}:
 $([1, 2])
 $([3, 4])

Utilities

PAndQ.value — Function

value(p)

If p is logically equivalent to a Constant, return that value wrapped in Some. Otherwise, return nothing.

Values wrapped in Some can be unwrapped using the something function.

Examples

julia> @atomize value(p)

julia> @atomize value($1)
Some(1)

julia> @atomize something(value($2))
2

PAndQ.atoms — Function

atoms(p)

Return an iterator of each Atom contained in p.

Examples

julia> @atomize collect(atoms(p ∧ q))
2-element Vector{PAndQ.Variable}:
 p
 q

PAndQ.operators — Function

operators(p)

Return an iterator of each operator contained in p.

Examples

julia> @atomize collect(operators(¬p))
1-element Vector{Operator{:not}}:
 ¬

julia> @atomize collect(operators(¬p ∧ q))
3-element Vector{Operator}:
 ∧
 ¬
 𝒾

PAndQ.install_atomize_mode — Function

install_atomize_mode(;
    start_key = "\M-a", prompt_text = "atomize> ", prompt_color = :cyan,
kwargs...)

Install the atomize REPL mode, where input implicitly begins with @atomize.

Keyword arguments are passed to ReplMaker.initrepl. The default start keys are pressing both the [Meta] (also known as [Alt]) and [a] keys at the same time. The available prompt_colors are in Base.text_colors.

Normalization

PAndQ.normalize — Function

normalize(::Union{typeof(¬), typeof(∧), typeof(∨)}, p)

Convert the given proposition to negation, conjunction, or disjunction normal form depending on whether the first argument is not, and, or or, respectively.

Considering the syntax tree of a normalized proposition, each leaf node is a literal; either an Atom or it's negation. Propositions in negation normal form are expanded such that the syntax tree branches only contain the operators and and or. Conjunction and disjunction normal forms are negated normal forms that have been flattened by recursively distributing either the and or or operator over the other. In other words, a collection of literals is a clause and a proposition in conjunctive or disjunctive normal form is a conjunction of disjunctive clauses or a disjunction of conjunctive clauses, respectively.

Conjunctive and disjunctive, but not negation, normal forms are called canonical. Distributing an operator during conversion increases the size of the syntax tree exponentially. Therefore, it is not possible to compute the canonical form for sufficiently large propositions. Use the tseytin transformation to find a proposition in conjunctive normal form which is_equisatisfiable.

Operations between canonical propositions return another canonical proposition, while operations between canonical and non-canonical propositions return a non-canonical proposition. It is performant to apply the not operator to a proposition in canonical normal form and the and or or operator to two propositions in conjunction or disjunction normal form, respectively. It is not performant to convert a proposition between conjunction and disjunction normal form. Therefore, it is typically more performant to first perform operations on non-canonical propositions before converting them to a canonical form.

Examples

julia> @atomize normalize(∧, ¬(p ∨ q))
(¬p) ∧ (¬q)

julia> @atomize normalize(∨, p ↔ q)
(¬q ∧ ¬p) ∨ (q ∧ p)

PAndQ.tseytin — Function

tseytin(p)

Apply the Tseytin transformation to the given proposition.

The transformed proposition is normalized to conjunction form, contains introduced Variables, and is_equisatisfiable to p. The valuations of the transformed proposition that result in a true interpretation are a subset of the same for p.

Examples

julia> is_equisatisfiable(⊤, tseytin(⊤))
true

julia> @atomize is_equisatisfiable(p, tseytin(p))
true

julia> is_equisatisfiable(⊥, tseytin(⊥))
true

PAndQ.dimacs — Function

dimacs(io = stdout, p)

Examples

julia> @atomize dimacs(p ↔ q)
p cnf 2 2
1 -2 0
-1 2 0

julia> @atomize dimacs(String, p ↔ q)
"p cnf 2 2\n1 -2 0\n-1 2 0\n"