Propositions

Tip

Propositions can be converted into different, but logically equivalent forms (see also ==). For example, ⊥ == Tree(⊥) == Clause(⊥) == Normal(⊥). However, not all forms are Expressively complete. Otherwise, the conversion may throw an exception. For example, there is no way to represent Literal(⊥).

Proposition
├─ Atom
└─ Compound
   ├─ Clause
   ├─ Expressive
   │  ├─ Normal
   │  └─ Tree
   └─ Literal

Abstract Types

PAndQ.Proposition — Type

Proposition

The set of well-formed logical formulae.

Supertype of Atom and Compound.

PAndQ.Compound — Type

Compound{LO} <: Proposition

A proposition composed from connecting Atomic propositions with LogicalOperators.

Subtype of Proposition. Supertype of Literal, Clause, and Expressive.

PAndQ.Expressive — Type

Expressive{LO} <: Compound{LO}

A proposition that is expressively complete.

Subtype of Compound. Supertype of Tree and Normal.

Concrete Types

PAndQ.Atom — Type

Atom{T} <: Proposition
Atom(::T)

A proposition with no deeper propositional structure.

Tip

Define pretty-printing for an instance of Atom{T} by overloading show(io::IO, ::MIME"text/plain", p::Atom{T}).

Tip

Use @atoms and @p as shortcuts to define atoms or instantiate them inline, respectively.

Subtype of Proposition.

Examples

julia> Atom(:p)
p

julia> Atom("Logic is fun")
Atom("Logic is fun")

PAndQ.Literal — Type

Literal{UO <: UnaryOperator, T} <: Compound{UO}
Literal(::UO, ::Atom{T})
Literal(::LiteralProposition)

A proposition represented by an atomic formula or its negation.

Subtype of Compound. See also UnaryOperator, Atom, and LiteralProposition.

Examples

julia> r = @p ¬p
¬p

julia> ¬r
p

PAndQ.Tree — Type

Tree{LO <: LogicalOperator, AT <: Union{Atom, Tree}} <: Expressive{LO}
Tree(::NullaryOperator, ::Atom)
Tree(::LogicalOperator, ::Tree...)
Tree(::Proposition)

A proposition represented by an abstract syntax tree.

Subtype of Expressive. See also LogicalOperator.

Examples

julia> @p r = p ⊻ q
p ⊻ q

julia> @p ¬r → s
(p ↔ q) → s

PAndQ.Clause — Type

Clause{AO <: AndOr, L <: Literal} <: Compound{AO}
Clause(::AO, ps = Literal[])
Clause(::AO, p::Proposition)
Clause(::Union{NullaryOperator, LiteralProposition})

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty clause is logically equivalent to the neutral element of it's binary operator.

Subtype of Compound. See also AndOr, Literal, NullaryOperator, and LiteralProposition.

Examples

julia> Clause(and)
⊤

julia> @p Clause(p)
p

julia> @p Clause(or, [¬p, q])
¬p ∨ q

PAndQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Expressive{AO}
Normal(::typeof(and), ps = Clause{typeof(or)}[])
Normal(::typeof(or), ps = Clause{typeof(and)}[])
Normal(::AO, ::Proposition)
Normal(::Union{NullaryOperator, Proposition})

A proposition represented in conjunctive or disjunctive normal form.

Info

An empty normal form is logically equivalent to the neutral element of it's binary operator.

Subtype of Expressive. See also AndOr, Clause, NullaryOperator, and Proposition.

Examples

julia> s = @p Normal(and, [Clause(or, [p, q]), Clause(or, ¬r)])
(p ∨ q) ∧ (¬r)

julia> ¬s
(¬p ∧ ¬q) ∨ (r)

Macros

PAndQ.@atoms — Macro

@atoms(ps...)

Instantiate and define Atoms with symbols and return a vector containing them.

Info

Atoms are defined in the global scope as constants.

Examples

julia> @atoms p q
2-element Vector{Atom{Symbol}}:
 p
 q

julia> p
p

julia> q
q

PAndQ.@p — Macro

@p(expression)

Instantiates each undefined variable (ignoring variable assignment and keyword arguments) as an Atom{Symbol}.

Examples

julia> @p x = p ∧ q
p ∧ q

julia> @p x → r
(p ∧ q) → r

PAndQ.@p_str — Macro

@p_str(x)

Examples

julia> p = @p_str("x")
x

julia> p"p ∧ q, Clause(and)"
(Tree(and, Tree(identity, Atom(:x)), Tree(identity, Atom(:q))), Clause(and, []))

Utilities

PAndQ.atoms — Function

atoms(p)

Returns an iterator of Atoms contained in p.

Examples

julia> @p collect(atoms(¬p))
1-element Vector{Atom{Symbol}}:
 p

julia> @p collect(atoms(p ∧ q))
2-element Vector{Atom{Symbol}}:
 p
 q

PAndQ.operators — Function

operators(p)

Returns an iterator of LogicalOperators contained in p.

Each node of p is iterated over in a pre-order tree traversal.

Examples

julia> @p collect(operators(¬p))
1-element Vector{typeof(not)}:
 not (generic function with 19 methods)

julia> @p collect(operators(¬p ∧ q))
3-element Vector{Function}:
 and (generic function with 23 methods)
 not (generic function with 19 methods)
 identity (generic function with 1 method)