Propositions

Tip

Propositions can be converted into different, but logically equivalent forms (see also ==). For example, ⊥ == Tree(⊥) == Clause(⊥) == Normal(⊥). However, not all forms are Expressively complete. Otherwise, the conversion may throw an exception. For example, there is no way to represent Literal(⊥).

Proposition
├─ Atom
└─ Compound
   ├─ Clause
   ├─ Expressive
   │  ├─ Normal
   │  └─ Tree
   └─ Literal

PAQ.Proposition — Type

Proposition

The set of well-formed logical formulae.

Supertype of Atom and Compound.

PAQ.Compound — Type

Compound <: Proposition

A proposition composed from connecting Atomicpropositions with LogicalOperators.

Subtype of Proposition. Supertype of Literal, Clause, and Expressive.

PAQ.Expressive — Type

Expressive <: Compound

A proposition that is expressively complete.

Subtype of Compound. Supertype of Tree and Normal.

PAQ.Atom — Type

Atom{T} <: Proposition
Atom(::T = :_)
Atom(::AtomicProposition)

A proposition with no deeper propositional structure.

Tip

Define pretty-printing for an instance of Atom{T} by overloading show(io::IO, p::Atom{T}).

Tip

Use @atoms or @p as a shortcut to define atoms or instantiate them inline, respectively.

Info

The default parameter :_ represents an Atom with an unspecified statement. For example, Tree(⊥) returns Tree(and(Atom(:_), not(Atom(:_))), which pretty-prints as _ ∧ ¬_. The underscore is a special case; it is not idiomatic to use for most purposes.

Subtype of Proposition. See also AtomicProposition.

Examples

julia> Atom(:p)
Atom:
 p

julia> Atom("Logic is fun")
Atom:
 "Logic is fun"

PAQ.Literal — Type

Literal{UO <: UnaryOperator, T} <: Compound
Literal(::UO, ::Atom)
Literal(::LiteralProposition)

A proposition represented by an atomic formula or its negation.

Subtype of Compound. See also UnaryOperator, Atom, and LiteralProposition.

Examples

julia> r = @p ¬p
Literal:
 ¬p

julia> ¬r
Atom:
 p

PAQ.Tree — Type

Tree{
    LO <: LogicalOperator,
    P <: Union{Tuple{Proposition}, Tuple{Proposition, Proposition}}
} <: Expressive
Tree(::UnaryOperator, ::Atom)
Tree(::BinaryOperator, ::Tree, ::Tree)
Tree(x)

A proposition represented by an abstract syntax tree.

Subtype of Expressive.

Examples

julia> @p r = p ⊻ q
Tree:
 p ⊻ q

julia> @p ¬r → s
Tree:
 (p ↔ q) → s

PAQ.Clause — Type

Clause{AO <: AndOr, L <: Literal} <: Compound
Clause(::AO, ::Vector = Literal[])
Clause(::AO, ps...)

A proposition represented as either a conjunction or disjunction of literals.

Info

An empty clause is logically equivalent to the identity element of it's binary operator.

See also Literal. Subtype of Compound.

Examples

julia> Clause(and)
Clause:
 ⊥

julia> @p Clause(and, p, q)
Clause:
 p ∧ q

julia> @p Clause(or, [¬p, q])
Clause:
 ¬p ∨ q

PAQ.Normal — Type

Normal{AO <: AndOr, C <: Clause} <: Expressive
Normal(::A, ::Vector{C} = C[]) where {A <: typeof(and), C <: Clause{typeof(or)}}
Normal(::O, ::Vector{C} = C[]) where {O <: typeof(or), C <: Clause{typeof(and)}}
Normal(::AO, ps...)

A proposition represented in conjunctive or disjunctive normal form.

Info

An empty normal form is logically equivalent to the identity element of it's binary operator.

Subtype of Expressive.

Examples

julia> s = @p Normal(and, Clause(or, p, q), Clause(or, ¬r))
Normal:
 (p ∨ q) ∧ (¬r)

julia> ¬s
Normal:
 (¬p ∧ ¬q) ∨ (r)