Semantics

Base.:== — Function

p == q
==(p, q)

Returns a boolean indicating whether p and q are logically equivalent.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it. Use == and === to test for equivalence and identity, respectively.

See also Proposition.

Examples

julia> @p p == ¬p
false

julia> @p ¬(p ⊻ q) == (p → q) ∧ (p ← q)
true

julia> @p ¬(p ⊻ q) === (p → q) ∧ (p ← q)
false

PAQ.interpret — Function

interpret(p::Union{NullaryOperator, Proposition}, valuation...)

Replaces each Atom in p with its truth value in valuation, then simplifies.

valuation is either a Dict or a set that can construct one that maps from atoms to their respective truth values.

Calling p with an incomplete mapping will partially interpret it.

See also tautology and [contradiction].

Examples

julia> @p interpret(¬p, p => ⊤)
contradiction (generic function with 1 method)

julia> @p p = Clause(and, q, r, s)
Clause:
 q ∧ r ∧ s

julia> @p interpret(p, q => ⊤, r => ⊤)
Clause:
 s

PAQ.interpretations — Function

interpretations(p, valuations = valuations(p))

Return an iterator of values given by interpreting p by each valuation.

See also valuations.

Examples

julia> @p collect(interpretations(p))
2-element Vector{Function}:
 tautology (generic function with 1 method)
 contradiction (generic function with 1 method)

julia> @p collect(interpretations(p → q, [p => ⊤]))
1-element Vector{Normal{typeof(and), Clause{typeof(or), Literal{typeof(identity), Symbol}}}}:
 (q)

PAQ.valuations — Function

valuations(atoms)
valuations(::Proposition)

Return an iterator containing every possible valuation of the Atoms.

A valuation is a vector of Pairs which map from an atom to a truth value.

Examples

julia> @p collect(valuations([p]))
2-element Vector{Vector}:
 Pair{Atom{Symbol}, typeof(tautology)}[p => PAQ.tautology]
 Pair{Atom{Symbol}, typeof(contradiction)}[p => PAQ.contradiction]

julia> @p collect(valuations([p, q]))
4-element Vector{Vector}:
 Pair{Atom{Symbol}, typeof(tautology)}[p => PAQ.tautology, q => PAQ.tautology]
 Pair{Atom{Symbol}}[p => PAQ.contradiction, q => PAQ.tautology]
 Pair{Atom{Symbol}}[p => PAQ.tautology, q => PAQ.contradiction]
 Pair{Atom{Symbol}, typeof(contradiction)}[p => PAQ.contradiction, q => PAQ.contradiction]

PAQ.solve — Function

solve(p, truth_value = ⊤)

Return a vector containing all interpretations such that interpret(p, interpretation) == truth_value.

Examples

julia> @p collect(solve(p))
1-element Vector{Vector{Pair{Atom{Symbol}, typeof(tautology)}}}:
 [p => PAQ.tautology]

julia> @p collect(solve(p ⊻ q, ⊥))
2-element Vector{Vector}:
 Pair{Atom{Symbol}, typeof(tautology)}[p => PAQ.tautology, q => PAQ.tautology]
 Pair{Atom{Symbol}, typeof(contradiction)}[p => PAQ.contradiction, q => PAQ.contradiction]

PAQ.left_identity — Function

left_identity(::LogicalOperator)

Return the corresponding identity element or nothing if it does not exist. The identity element is either tautology or contradiction.

Examples

julia> left_identity(or)
contradiction (generic function with 1 method)

julia> left_identity(imply)
tautology (generic function with 1 method)

PAQ.right_identity — Function

right_identity(::LogicalOperator)

Return the corresponding identity element or nothing if it does not exist. The identity element is either tautology or contradiction.

Examples

julia> right_identity(or)
contradiction (generic function with 1 method)

julia> right_identity(converse_imply)
tautology (generic function with 1 method)

PAQ.converse — Function

converse(::LogicalOperator)

Returns the LogicalOperator that is the converse of the given boolean operator.

Examples

julia> converse(and)
and (generic function with 23 methods)

julia> @p and(p, q) == converse(and)(q, p)
true

julia> converse(imply)
converse_imply (generic function with 7 methods)

julia> @p imply(p, q) == converse(imply)(q, p)
true

PAQ.dual — Function

dual(::LogicalOperator)

Returns the LogicalOperator that is the dual of the given boolean operator.

Examples

julia> dual(and)
or (generic function with 19 methods)

julia> @p and(p, q) == not(dual(and)(not(p), not(q)))
true

julia> dual(imply)
not_converse_imply (generic function with 6 methods)

julia> @p imply(p, q) == not(dual(imply)(not(p), not(q)))
true

PAQ.is_tautology — Function

is_tautology(p)

Returns a boolean on whether p is a tautology.

Examples

julia> is_tautology(⊤)
true

julia> @p is_tautology(p)
false

julia> @p is_tautology(¬(p ∧ ¬p))
true

PAQ.is_contradiction — Function

is_contradiction(p)

Returns a boolean on whether p is a contradiction.

Examples

julia> is_contradiction(⊥)
true

julia> @p is_contradiction(p)
false

julia> @p is_contradiction(p ∧ ¬p)
true

PAQ.is_truth — Function

is_truth(p)

Returns a boolean on whether p is a truth value (either a tautology or contradiction).

See also Proposition.

Examples

julia> is_truth(⊤)
true

julia> @p is_truth(p ∧ ¬p)
true

julia> @p is_truth(p)
false

julia> @p is_truth(p ∧ q)
false

PAQ.is_contingency — Function

is_contingency(p)

Returns a boolean on whether p is a contingency (neither a tautology or contradiction).

See also Proposition.

Examples

julia> is_contingency(⊤)
false

julia> @p is_contingency(p ∧ ¬p)
false

julia> @p is_contingency(p)
true

julia> @p is_contingency(p ∧ q)
true

PAQ.is_satisfiable — Function

is_satisfiable(p)

Returns a boolean on whether p is satisfiable (not a contradiction).

See also Proposition.

Examples

julia> is_satisfiable(⊤)
true

julia> @p is_satisfiable(p ∧ ¬p)
false

julia> @p is_satisfiable(p)
true

julia> @p is_satisfiable(p ∧ q)
true

PAQ.is_falsifiable — Function

is_falsifiable(p)

Returns a boolean on whether p is falsifiable (not a tautology).

See also Proposition.

Examples

julia> is_falsifiable(⊥)
true

julia> @p is_falsifiable(¬(p ∧ ¬p))
false

julia> @p is_falsifiable(p)
true

julia> @p is_falsifiable(p ∧ q)
true