Semantics

Properties

PAndQ.dual — Function

dual(::LogicalOperator)

Returns the LogicalOperator that is the dual of the given boolean operator.

Examples

julia> dual(and)
or (generic function with 19 methods)

julia> @p and(p, q) == not(dual(and)(not(p), not(q)))
true

julia> dual(imply)
not_converse_imply (generic function with 6 methods)

julia> @p imply(p, q) == not(dual(imply)(not(p), not(q)))
true

PAndQ.converse — Function

converse(::LogicalOperator)

Returns the LogicalOperator that is the converse of the given boolean operator.

Examples

julia> converse(and)
and (generic function with 23 methods)

julia> @p and(p, q) == converse(and)(q, p)
true

julia> converse(imply)
converse_imply (generic function with 7 methods)

julia> @p imply(p, q) == converse(imply)(q, p)
true

PAndQ.left_neutrals — Function

left_neutrals(::LogicalOperator)

Return the corresponding left identity elements of the operator. The identity elements can be tautology, contradiction, neither (empty set), or both.

Examples

julia> left_neutrals(or)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.contradiction

julia> left_neutrals(imply)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.tautology

julia> left_neutrals(nor)
Set{Union{typeof(contradiction), typeof(tautology)}}()

PAndQ.right_neutrals — Function

right_neutrals(::LogicalOperator)

Return the corresponding right identity elements of the operator. The identity elements can be tautology, contradiction, neither (empty set), or both.

Examples

julia> right_neutrals(or)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.contradiction

julia> right_neutrals(converse_imply)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.tautology

Truths

PAndQ.valuations — Function

valuations(as, n = length(unique(atoms(as))))
valuations(::Proposition)

Return an iterator of every possible valuation of the Atoms.

Examples

julia> @p collect(valuations([p]))
2-element Vector{Vector}:
 Pair{Atom{Symbol}, typeof(tautology)}[Atom(:p) => PAndQ.tautology]
 Pair{Atom{Symbol}, typeof(contradiction)}[Atom(:p) => PAndQ.contradiction]

julia> @p collect(valuations([p, q]))
4-element Vector{Vector}:
 Pair{Atom{Symbol}, typeof(tautology)}[Atom(:p) => PAndQ.tautology, Atom(:q) => PAndQ.tautology]
 Pair{Atom{Symbol}}[Atom(:p) => PAndQ.contradiction, Atom(:q) => PAndQ.tautology]
 Pair{Atom{Symbol}}[Atom(:p) => PAndQ.tautology, Atom(:q) => PAndQ.contradiction]
 Pair{Atom{Symbol}, typeof(contradiction)}[Atom(:p) => PAndQ.contradiction, Atom(:q) => PAndQ.contradiction]

PAndQ.interpret — Function

interpret(valuation, ::Proposition)

Replaces each Atom p in the given Proposition with valuation(p), then simplifies.

Calling p with an incomplete mapping will partially interpret it.

See also tautology and contradiction.

Examples

julia> @p interpret(a -> ⊤, ¬p)
contradiction (generic function with 1 method)

julia> @p interpret(a -> get(Dict(p => ⊤), a, a), p ∧ q)
(q)

(::Proposition)(valuation)
(::Proposition)(valuation...)

Equivalent to interpret(a -> get(Dict(valuation), a, a), p).

See also Proposition.

Examples

julia> @p (¬p)(p => ⊤)
contradiction (generic function with 1 method)

julia> @p p = Clause(and, [q, r, s])
q ∧ r ∧ s

julia> @p p(q => ⊤, r => ⊤)
s

PAndQ.interpretations — Function

interpretations(p, valuations = valuations(p))

Return an iterator of truth values given by interpreting p with each valuation.

Examples

julia> @p collect(interpretations(p))
2-element Vector{Function}:
 tautology (generic function with 1 method)
 contradiction (generic function with 1 method)

julia> @p collect(interpretations(p → q, [p => ⊤]))
1-element Vector{Normal{typeof(and), Clause{typeof(or)}}}:
 (q)

PAndQ.solve — Function

solve(p)

Return a vector containing all interpretations such that interpret(p, valuation) == tautology.

Examples

julia> @p collect(solve(p))
1-element Vector{Vector{Pair{Atom{Symbol}, typeof(tautology)}}}:
 [Atom(:p) => PAndQ.tautology]

julia> @p collect(solve(p ⊻ q))
2-element Vector{Vector{Pair{Atom{Symbol}}}}:
 [Atom(:p) => PAndQ.contradiction, Atom(:q) => PAndQ.tautology]
 [Atom(:p) => PAndQ.tautology, Atom(:q) => PAndQ.contradiction]

Predicates

Base.:== — Function

==(::Union{NullaryOperator, Proposition}, ::Union{NullaryOperator, Proposition})
p == q

Returns a boolean indicating whether p and q are logically equivalent.

Atoms are equivalent if and only if their statements are equivalent.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it. Use == and === to test for equivalence and identity, respectively.

See also NullaryOperator and Proposition.

Examples

julia> @p p == ¬p
false

julia> @p ¬(p ⊻ q) == (p → q) ∧ (p ← q)
true

julia> @p ¬(p ⊻ q) === (p → q) ∧ (p ← q)
false

PAndQ.is_tautology — Function

is_tautology(p)

Returns a boolean on whether p is a tautology.

Examples

julia> is_tautology(⊤)
true

julia> @p is_tautology(p)
false

julia> @p is_tautology(¬(p ∧ ¬p))
true

PAndQ.is_contradiction — Function

is_contradiction(p)

Returns a boolean on whether p is a contradiction.

Examples

julia> is_contradiction(⊥)
true

julia> @p is_contradiction(p)
false

julia> @p is_contradiction(p ∧ ¬p)
true

PAndQ.is_truth — Function

is_truth(p)

Returns a boolean on whether p is a truth value (either a tautology or contradiction).

See also Proposition.

Examples

julia> is_truth(⊤)
true

julia> @p is_truth(p ∧ ¬p)
true

julia> @p is_truth(p)
false

julia> @p is_truth(p ∧ q)
false

PAndQ.is_contingency — Function

is_contingency(p)

Returns a boolean on whether p is a contingency (neither a tautology or contradiction).

See also Proposition.

Examples

julia> is_contingency(⊤)
false

julia> @p is_contingency(p ∧ ¬p)
false

julia> @p is_contingency(p)
true

julia> @p is_contingency(p ∧ q)
true

PAndQ.is_satisfiable — Function

is_satisfiable(p)

Returns a boolean on whether p is satisfiable (not a contradiction).

See also Proposition.

Examples

julia> is_satisfiable(⊤)
true

julia> @p is_satisfiable(p ∧ ¬p)
false

julia> @p is_satisfiable(p)
true

julia> @p is_satisfiable(p ∧ q)
true

PAndQ.is_falsifiable — Function

is_falsifiable(p)

Returns a boolean on whether p is falsifiable (not a tautology).

See also Proposition.

Examples

julia> is_falsifiable(⊥)
true

julia> @p is_falsifiable(¬(p ∧ ¬p))
false

julia> @p is_falsifiable(p)
true

julia> @p is_falsifiable(p ∧ q)
true

Utilities

Base.convert — Function

convert(::Type{<:Proposition}, ::Union{NullaryOperator, Proposition})