Semantics

PAQ.interpret — Function

interpret(valuation, ϕ::Language)

Given a valuation function that maps from the Primitive propositions in ϕ to their respective Truth values, assign a truth value to ϕ.

See also Language.

Base.:== — Function

p == q
==(p::Language, q::Language)
isequal(p::Language, q::Language)

Returns a boolean indicating whether p and q are logically equivalent.

See also Language.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it.

Examples

julia> p == ¬p
false

julia> p ∨ q == ¬(¬q ∧ ¬p)
true

julia> isequal((p → q) ∧ (p ← q), ¬(p ⊻ q))
true

PAQ.is_tautology — Function

is_tautology(p::Language)

Returns a boolean on whether the given proposition is a tautology.

This function is equivalent to p == ⊤.

See also Language and ==.

Examples

julia> is_tautology(⊤)
true

julia> is_tautology(p)
false

julia> is_tautology(¬(p ∧ ¬p))
true

PAQ.is_contradiction — Function

is_contradiction(p::Language)

Returns a boolean on whether the given proposition is a contradiction.

This function is equivalent to p == ⊥.

See also Language and ==.

Examples

julia> is_contradiction(⊥)
true

julia> is_contradiction(p)
false

julia> is_contradiction(p ∧ ¬p)
true

PAQ.is_truth — Function

is_truth(p::Language)

Returns a boolean on whether the given proposition is a Truth (either a tautology or contradiction).

See also Language.

Examples

julia> is_truth(⊤)
true

julia> is_truth(p ∧ ¬p)
true

julia> is_truth(p)
false

julia> is_truth(p ∧ q)
false

PAQ.is_contingency — Function

is_contingency(p::Language)

Returns a boolean on whether the given proposition is a contingency (neither a tautology or contradiction).

See also Language.

Examples

julia> is_contingency(⊤)
false

julia> is_contingency(p ∧ ¬p)
false

julia> is_contingency(p)
true

julia> is_contingency(p ∧ q)
true

PAQ.is_satisfiable — Function

is_satisfiable(p::Language)

Returns a boolean on whether the given proposition is satisfiable (not a contradiction).

See also Language.

Examples

julia> is_satisfiable(⊤)
true

julia> is_satisfiable(p ∧ ¬p)
false

julia> is_satisfiable(p)
true

julia> is_satisfiable(p ∧ q)
true

PAQ.is_falsifiable — Function

is_falsifiable(p::Language)

Returns a boolean on whether the given proposition is falsifiable (not a is_tautology).

See also Language.

Examples

julia> is_falsifiable(⊥)
true

julia> is_falsifiable(¬(p ∧ ¬p))
false

julia> is_falsifiable(p)
true

julia> is_falsifiable(p ∧ q)
true

PAQ.@truth_table — Macro

@truth_table p
@truth_table(ps...)

Print a truth table for the given propositions.

The first row of the header is the expression representing that column's proposition, the second row indicates that expression's type, and the third row identifies the statements for Primitive propositions.

Info

If a variable contains a primitive, there is no expression to label that primitive. As such, the first row in the header will be blank. However, the identifying statement is still known and will be displayed in the third row. Use get_primitives to resolve this uncertainty.

Logically equivalent propositions will be placed in the same column with their expressions in the header seperated by a comma.

In this context, ⊤ and ⊥ can be interpreted as true and false, respectively.

See also Language.

Examples

julia> @truth_table p ∧ q p → q
┌───────────┬───────────┬───────────────┬───────────────┐
│ p         │ q         │ p ∧ q         │ p → q         │
│ Primitive │ Primitive │ Propositional │ Propositional │
│ "p"       │ "q"       │               │               │
├───────────┼───────────┼───────────────┼───────────────┤
│ ⊤         │ ⊤         │ ⊤             │ ⊤             │
│ ⊤         │ ⊥         │ ⊥             │ ⊥             │
├───────────┼───────────┼───────────────┼───────────────┤
│ ⊥         │ ⊤         │ ⊥             │ ⊤             │
│ ⊥         │ ⊥         │ ⊥             │ ⊤             │
└───────────┴───────────┴───────────────┴───────────────┘