Semantics

PAQ.interpret — Function

interpret(p, valuation...)

Assign a truth value value to p.

Let p be a Proposition. Let valuation be a function, callable object, dictionary, or any number of Pairs that map from atomic propositions in p to their respective truth values.

Calling p with an incomplete mapping will partially interpret p. This returns a Proposition of the same type as p?*? that is independent from every Atoms in valuation.

Warning

If valuation does not return a Truth or errors,

Examples

Base.:== — Function

p == q
==(p, q)
isequal(p, q)

Returns a boolean indicating whether p and q are logically equivalent.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it. Use == and === to test for equivalence and identity, respectively.

See also Proposition.

Examples

julia> @p p == ¬p
false

julia> @p (p → q) ∧ (p ← q) == ¬(p ⊻ q)
true

julia> @p (p → q) ∧ (p ← q) === ¬(p ⊻ q)
false

PAQ.is_tautology — Function

is_tautology(p)

Returns a boolean on whether p is a tautology.

Examples

julia> is_tautology(⊤)
true

julia> @p is_tautology(p)
false

julia> @p is_tautology(¬(p ∧ ¬p))
true

PAQ.is_contradiction — Function

is_contradiction(p)

Returns a boolean on whether p is a contradiction.

Examples

julia> is_contradiction(⊥)
true

julia> @p is_contradiction(p)
false

julia> @p is_contradiction(p ∧ ¬p)
true

PAQ.is_truth — Function

is_truth(p)

Returns a boolean on whether p is a truth value (either a tautology or contradiction).

See also Proposition.

Examples

julia> is_truth(⊤)
true

julia> @p is_truth(p ∧ ¬p)
true

julia> @p is_truth(p)
false

julia> @p is_truth(p ∧ q)
false

PAQ.is_contingency — Function

is_contingency(p)

Returns a boolean on whether p is a contingency (neither a tautology or contradiction).

See also Proposition.

Examples

julia> is_contingency(⊤)
false

julia> @p is_contingency(p ∧ ¬p)
false

julia> @p is_contingency(p)
true

julia> @p is_contingency(p ∧ q)
true

PAQ.is_satisfiable — Function

is_satisfiable(p)

Returns a boolean on whether p is satisfiable (not a contradiction).

See also Proposition.

Examples

julia> is_satisfiable(⊤)
true

julia> @p is_satisfiable(p ∧ ¬p)
false

julia> @p is_satisfiable(p)
true

julia> @p is_satisfiable(p ∧ q)
true

PAQ.is_falsifiable — Function

is_falsifiable(p)

Returns a boolean on whether p is falsifiable (not a tautology).

See also Proposition.

Examples

julia> is_falsifiable(⊥)
true

julia> @p is_falsifiable(¬(p ∧ ¬p))
false

julia> @p is_falsifiable(p)
true

julia> @p is_falsifiable(p ∧ q)
true

PAQ.solve — Function

solve(p)

Return a vector of every valid interpretation of p.

Examples

julia> @p solve(p ⊻ q)
2-element Vector{Vector{Pair{Atom{Symbol}}}}:
 [p => ⊥, q => ⊤]
 [p => ⊤, q => ⊥]

PAQ.converse — Function

converse(::BooleanOperator)

Returns the BooleanOperator that is the converse of the given boolean operator.

Examples

julia> converse(and)
and (generic function with 20 methods)

julia> @p and(p, q) == converse(and)(q, p)
true

julia> converse(imply)
converse_imply (generic function with 9 methods)

julia> @p imply(p, q) == converse(imply)(q, p)
true

PAQ.dual — Function

dual(::BooleanOperator)

Returns the BooleanOperator that is the dual of the given boolean operator.

Examples

julia> dual(and)
or (generic function with 17 methods)

julia> @p and(p, q) == not(dual(and)(not(p), not(q)))
true

julia> dual(imply)
not_converse_imply (generic function with 9 methods)

julia> @p imply(p, q) == not(dual(imply)(not(p), not(q)))
true