Semantics

Properties

PAndQ.dual — Function

dual(::LogicalOperator)

Returns the LogicalOperator that is the dual of the given boolean operator.

Examples

julia> dual(and)
or (generic function with 18 methods)

julia> @atomize and(p, q) == not(dual(and)(not(p), not(q)))
true

julia> dual(imply)
not_converse_imply (generic function with 3 methods)

julia> @atomize imply(p, q) == not(dual(imply)(not(p), not(q)))
true

PAndQ.converse — Function

converse(::LogicalOperator)

Returns the LogicalOperator that is the converse of the given boolean operator.

Examples

julia> converse(and)
and (generic function with 18 methods)

julia> @atomize and(p, q) == converse(and)(q, p)
true

julia> converse(imply)
converse_imply (generic function with 3 methods)

julia> @atomize imply(p, q) == converse(imply)(q, p)
true

PAndQ.left_neutrals — Function

left_neutrals(::LogicalOperator)

Return the corresponding left identity elements of the operator. The identity elements can be tautology, contradiction, neither (empty set), or both.

Examples

julia> left_neutrals(or)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.contradiction

julia> left_neutrals(imply)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.tautology

julia> left_neutrals(nor)
Set{Union{typeof(contradiction), typeof(tautology)}}()

PAndQ.right_neutrals — Function

right_neutrals(::LogicalOperator)

Return the corresponding right identity elements of the operator. The identity elements can be tautology, contradiction, neither (empty set), or both.

Examples

julia> right_neutrals(or)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.contradiction

julia> right_neutrals(converse_imply)
Set{Union{typeof(contradiction), typeof(tautology)}} with 1 element:
  PAndQ.tautology

Truths

PAndQ.valuations — Function

valuations(atoms)
valuations(::Proposition)

Return an iterator of every possible valuation of Atoms.

Examples

julia> @atomize collect(valuations(p))
2-element Vector{Vector{Pair{Variable, Bool}}}:
 [Variable(:p) => 1]
 [Variable(:p) => 0]

julia> @atomize collect(valuations(p ∧ q))
2×2 Matrix{Vector{Pair{Variable, Bool}}}:
 [Variable(:p)=>1, Variable(:q)=>1]  [Variable(:p)=>1, Variable(:q)=>0]
 [Variable(:p)=>0, Variable(:q)=>1]  [Variable(:p)=>0, Variable(:q)=>0]

PAndQ.interpret — Function

interpret(valuation, ::Proposition)

Substitute each Atom in the given Proposition with values from the valuation.

The valuation can be a function or function-like object with the signature valuation(::Atom)::Union{Bool, NullaryOperator}, a Dict, or an iterable that can construct a Dict. No substitution is performed if an Atom from the Proposition is not one of the dictionary's keys.

See also NullaryOperator.

Examples

julia> @atomize interpret(a -> true, ¬p)
false

julia> @atomize interpret(p => true, p ∧ q)
q

PAndQ.interpretations — Function

interpretations(p, valuations = valuations(p))

Return an iterator of truth values given by interpreting p with each valuation.

Examples

julia> @atomize collect(interpretations(p))
2-element Vector{Bool}:
 1
 0

julia> @atomize collect(interpretations(p ⊻ q, [p => true]))
1-element Vector{Literal{typeof(not), Variable}}:
 ¬q

PAndQ.solve — Function

solve(p)

Return a vector containing all interpretations such that interpret(p, valuation) == ⊤.

See also interpret and tautology.

Examples

julia> @atomize collect(solve(p))
1-element Vector{Vector{Pair{Variable, Bool}}}:
 [Variable(:p) => 1]

julia> @atomize collect(solve(p ⊻ q))
2-element Vector{Vector{Pair{Variable, Bool}}}:
 [Variable(:p) => 0, Variable(:q) => 1]
 [Variable(:p) => 1, Variable(:q) => 0]

Predicates

Base.:== — Function

==(::Union{NullaryOperator, Proposition}, ::Union{NullaryOperator, Proposition})
p == q

Returns a boolean indicating whether p and q are logically equivalent.

Constants are equivalent if and only if their values are equivalent.

Info

The ≡ symbol is sometimes used to represent logical equivalence. However, Julia uses ≡ as an alias for the builtin function === which cannot have methods added to it. Use == and === to test for equivalence and identity, respectively.

See also NullaryOperator and Proposition.

Examples

julia> @atomize p == ¬p
false

julia> @atomize ¬(p ⊻ q) == (p → q) ∧ (p ← q)
true

julia> @atomize ¬(p ⊻ q) === (p → q) ∧ (p ← q)
false

PAndQ.is_tautology — Function

is_tautology(p)

Returns a boolean on whether p is a tautology.

Examples

julia> is_tautology(⊤)
true

julia> @atomize is_tautology(p)
false

julia> @atomize is_tautology(¬(p ∧ ¬p))
true

PAndQ.is_contradiction — Function

is_contradiction(p)

Returns a boolean on whether p is a contradiction.

Examples

julia> is_contradiction(⊥)
true

julia> @atomize is_contradiction(p)
false

julia> @atomize is_contradiction(p ∧ ¬p)
true

PAndQ.is_truth — Function

is_truth(p)

Returns a boolean on whether p is a truth value (either a tautology or contradiction).

See also Proposition.

Examples

julia> is_truth(⊤)
true

julia> @atomize is_truth(p ∧ ¬p)
true

julia> @atomize is_truth(p)
false

julia> @atomize is_truth(p ∧ q)
false

PAndQ.is_contingency — Function

is_contingency(p)

Returns a boolean on whether p is a contingency (neither a tautology or contradiction).

See also Proposition.

Examples

julia> is_contingency(⊤)
false

julia> @atomize is_contingency(p ∧ ¬p)
false

julia> @atomize is_contingency(p)
true

julia> @atomize is_contingency(p ∧ q)
true

PAndQ.is_satisfiable — Function

is_satisfiable(p)

Returns a boolean on whether p is satisfiable (not a contradiction).

See also Proposition.

Examples

julia> is_satisfiable(⊤)
true

julia> @atomize is_satisfiable(p ∧ ¬p)
false

julia> @atomize is_satisfiable(p)
true

julia> @atomize is_satisfiable(p ∧ q)
true

PAndQ.is_falsifiable — Function

is_falsifiable(p)

Returns a boolean on whether p is falsifiable (not a tautology).

See also Proposition.

Examples

julia> is_falsifiable(⊥)
true

julia> @atomize is_falsifiable(¬(p ∧ ¬p))
false

julia> @atomize is_falsifiable(p)
true

julia> @atomize is_falsifiable(p ∧ q)
true

Utilities

Base.convert — Function

convert(::Type{<:Proposition}, ::Union{NullaryOperator, Proposition})