Utility

PAndQ.arity — Function

arity(::LogicalOperator)

Returns the arity of the given LogicalOperator.

Examples

julia> arity(tautology)
0

julia> arity(not)
1

julia> arity(and)
2

PAndQ.@atoms — Macro

@atoms(ps...)

Instantiate and define Atoms with symbols and return a vector containing them.

Info

Atoms are defined in the global scope as constants.

Examples

julia> @atoms p q
2-element Vector{Atom{Symbol}}:
 p
 q

julia> p
Atom:
 p

julia> q
Atom:
 q

PAndQ.@p — Macro

@p(expression)

Instantiates each undefined variable (ignoring variable assignment and keyword arguments) as an Atom{Symbol}.

Examples

julia> @p x = p
Atom:
 p

julia> @p x ∧ q → r
Tree:
 (p ∧ q) → r

PAndQ.@p_str — Macro

@p_str(x)

Examples

julia> p = @p_str("x")
Atom:
 x

julia> p"p ∧ q, Clause(and)"
(x ∧ q, ⊤)

PAndQ.atoms — Function

atoms(::Proposition)

Returns a vector of unique Atoms contained in the given Proposition.

Warning

Some atoms may optimized out of an expression, such as in p ∧ ⊥ == ⊥.

Examples

julia> @p atoms(¬p)
1-element Vector{Atom{Symbol}}:
 p

julia> @p atoms(p ∧ q)
2-element Vector{Atom}:
 p
 q

PAndQ.conjunction — Function

⋀(ps)
conjunction(ps)

Equivalent to foldl(and, ps; init = ⊤).

⋀ can be typed by \bigwedge<tab>.

See also and.

Examples

julia> @p ⋀([p, q, r, s])
Tree:
 ((p ∧ q) ∧ r) ∧ s

PAndQ.disjunction — Function

⋁(ps)
disjunction(ps)

Equivalent to foldl(or, ps; init = ⊥).

⋁ can be typed by \bigvee<tab>.

See also or.

Examples

julia> @p ⋁([p, q, r, s])
Tree:
 ((p ∨ q) ∨ r) ∨ s