Utility

PAQ.primitives — Function

primitives(ps...)

Returns a vector of Primitive propositions contained in p.

Note that some primitives may optimized out of a statement, such as in p ∧ ⊥.

Examples

julia> primitives(p)
1-element Vector{Primitive}:
 Primitive("p")

julia> primitives(p ∧ q, r)
3-element Vector{Primitive}:
 Primitive("p")
 Primitive("q")
 Primitive("r")

PAQ.@primitive — Macro

@primitive(ps...)

Instantiates Primitive propositions.

Examples

julia> @primitive p q

julia> p
Primitive("p")

julia> q
Primitive("q")

PAQ.@truth_table — Macro

@truth_table(ps...)

Print a truth table for the given Compound propositions.

See also Primitive and Compound.

Examples

julia> @truth_table p∧q p→q
┌───────────┬───────────┬───────────────┬───────────────┐
│         p │         q │         p ∧ q │         p → q │
│ Primitive │ Primitive │ Propositional │ Propositional │
├───────────┼───────────┼───────────────┼───────────────┤
│         ⊤ │         ⊤ │             ⊤ │             ⊤ │
│         ⊤ │         ⊥ │             ⊥ │             ⊥ │
├───────────┼───────────┼───────────────┼───────────────┤
│         ⊥ │         ⊤ │             ⊤ │             ⊥ │
│         ⊥ │         ⊥ │             ⊤ │             ⊥ │
└───────────┴───────────┴───────────────┴───────────────┘

PAQ.interpret — Function

interpret(valuation, ϕ::Union{Primitive, Compound})

Given a valuation,

A valuation is a function that maps from the Primitive propositions in ϕ to their respective Truth values.

See also Primitive, Compound, and Truth.

julia> mapping = Dict(p => ⊥, q => ⊤);

julia> valuation = r -> mapping[r];

julia> interpret(valuation, p ∧ q)
⊥

julia> interpret(valuation, p → q)
⊤